% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_recomplexify.R
\name{account_missing}
\alias{account_missing}
\title{Function to account for missing subunits}
\usage{
account_missing(lr_res, env)
}
\arguments{
\item{lr_res}{LR results as obtained by `lr_pipe`}

\item{env}{environemnt to which the lr_res will be saved}
}
\description{
Function to account for missing subunits
}
\details{
this function applies loops over all complexes and assigns 0s to
   the stats of any complexes with a missing subunit. The loop is iteratively
   applied over the same `lr_res`, to provide a more controlled environment
   we pass the environment of recomplexify and modify `lr_res` in place.
}
