% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_omnipath.R
\name{assign_ligrecs}
\alias{assign_ligrecs}
\title{Helper function to obtain distinct transmitter and receiver lists
used in the resource comparison}
\usage{
assign_ligrecs(ligrec_list)
}
\arguments{
\item{ligrec_list}{e.g. ligrec$OmniPath}
}
\description{
Helper function to obtain distinct transmitter and receiver lists
used in the resource comparison
}
