% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_utils.R
\name{assign_lr_weights}
\alias{assign_lr_weights}
\title{Helper function to assign weights}
\usage{
assign_lr_weights(lrs, resource, entity = "ligand")
}
\arguments{
\item{lrs}{ligand_receptor tibble}

\item{resource}{resource}

\item{entity}{name of the entity}
}
\description{
Helper function to assign weights
}
