% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bench_robust.R
\name{bench_robust}
\alias{bench_robust}
\title{Benchmark Robustness Wrapper}
\usage{
bench_robust(subsampling, lr_call = NULL, seurat_object, ...)
}
\arguments{
\item{subsampling}{Vector of subsampling (e.g. 0.9, 0.7, 0.5)}

\item{lr_call}{Ligad Receptor Algorithm to call}

\item{seurat_object}{seurat object to subsample and use in the analyses}
}
\value{
LR results for subsampled Seurat objects
}
\description{
Benchmark Robustness Wrapper
}
