% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_descriptive.R
\name{binarize_resources}
\alias{binarize_resources}
\title{Helper Function to Convert a List of Resources into a binarized DF}
\usage{
binarize_resources(entity_list, cols)
}
\arguments{
\item{entity_list}{list resources with tibbles for entities
(i.e. interactions, transmitters, receivers)}
}
\value{
Returns a list of 1/0 tibbles for each entity where 1 is assigned
   to entities present in a given resource, 0s to absent ones
}
\description{
Helper Function to Convert a List of Resources into a binarized DF
}
