% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_roc.R
\name{calc_curve}
\alias{calc_curve}
\title{This function takes the elements of the lr_res column and calculates
robustness. The way that we assess robustness is by subsampling clusters.
Significant hits from the full data set are regarded as the ground truth
and results from subsampled data sets are evaluated against the ground truth.}
\usage{
calc_curve(
  df,
  ground,
  predictor_metric = "pvalue",
  downsampling = FALSE,
  times = 1000,
  curve = "ROC",
  seed = 1004
)
}
\arguments{
\item{downsampling}{logical flag indicating if the number of Negatives
should be downsampled to the number of Positives}

\item{times}{integer showing the number of downsampling}

\item{curve}{whether to return a Precision-Recall Curve ("PR") or ROC ("ROC")}

\item{seed}{An integer to set the RNG state for random number generation. Use
NULL for random number generation.}

\item{lr_form_res}{Formatted LR results with subsampling and the resources
used for each tool.}
}
\value{
tidy data frame with precision, recall, auc, n, cp, cn and coverage
   in the case of PR curve; or sensitivity and specificity, auc, n, cp, cn
   and coverage in the case of ROC.
}
\description{
This function takes the elements of the lr_res column and calculates
robustness. The way that we assess robustness is by subsampling clusters.
Significant hits from the full data set are regarded as the ground truth
and results from subsampled data sets are evaluated against the ground truth.
}
