% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_tensor.R
\name{calculate_gini}
\alias{calculate_gini}
\title{Function to calculate gini coefficients for source and target loadings}
\usage{
calculate_gini(loadings)
}
\arguments{
\item{loadings}{loadings for dimension of interest ('senders' or 'receivers')
formatted by `format_c2c_factors`}
}
\description{
Function to calculate gini coefficients for source and target loadings
}
\keyword{internal}
