% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellchat_pipe.R
\name{call_cellchat}
\alias{call_cellchat}
\title{Connectome x Omni Pipe function}
\usage{
call_cellchat(
  op_resource,
  seurat_object,
  .format = TRUE,
  exclude_anns = c("ecm-receptor", "cell_surface_ligand-receptor"),
  nboot = 100,
  assay = "SCT",
  .seed = 1004,
  .normalize = FALSE,
  .do_parallel = FALSE,
  ...
)
}
\arguments{
\item{op_resource}{OmniPath Intercell Resource DN}

\item{seurat_object}{Seurat object as input}

\item{.format}{bool whether to format output}

\item{exclude_anns}{Annotation criteria to be excluded}

\item{nboot}{number of bootstraps to calculate p-value}

\item{.normalize}{# bool whether to normalize non-normalized data with
internal func}
}
\value{
A DF of intercellular communication network
}
\description{
Connectome x Omni Pipe function
}
