% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytotalk.R
\name{call_cytotalk}
\alias{call_cytotalk}
\title{Compute cross-talk score from a Seurat Object}
\usage{
call_cytotalk(SeuratObject, op_resouce)
}
\arguments{
\item{SeuratObject}{Seurat object as input}

\item{op_resouce}{The resource of ligand-receptor interactions}
}
\value{
A data frame containing the resulting scores
}
\description{
This function executes all the required functions to extract and compute the
cross-talk scores as defined by the Cytotalk authors. It firsts transforms the SeuratObject
into a SingleCellExperiment object and then joins the results to the output of
the liana_pipe function.
}
