% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/italk_pipe.R
\name{call_italk}
\alias{call_italk}
\title{Run iTALK with OmniPath data [[DEPRECATED]]}
\usage{
call_italk(sce, op_resource, assay = "RNA", .format = TRUE, .DE = TRUE, ...)
}
\arguments{
\item{sce}{Seurat object or SingleCellExperiment as input}

\item{op_resource}{OmniPath Intercell Resource DN}

\item{assay}{assay to use from Seurat object}

\item{.format}{bool: whether to format output}

\item{.DE}{bool: whether to use DE (TRUE) or highlyVarGenes (FALSE)}
}
\value{
An unfiltered iTALK df sorted by relevance
}
\description{
Run iTALK with OmniPath data [[DEPRECATED]]
}
\details{
Stats:
Ligand and Receptor Expressions (and P-values if ran with .DE==TRUE)
Dot params are inherited from Seurat::FindAllMarkers, if .deg = TRUE

In this case, we use the product of the logFC rather than thresholding, as in
the original implementation.
}
