% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NATMI_pipe.R
\name{call_natmi}
\alias{call_natmi}
\title{Call NATMI Pipeline from R with OmniPath}
\usage{
call_natmi(
  omni_resources,
  seurat_object = NULL,
  wd_path = "/home/dbdimitrov/Repos/ligrec_decoupleR",
  omnidbs_path = "~/Repos/ligrec_decoupleR/input/omnipath_NATMI",
  natmi_path = "~/Repos/NATMI",
  em_path = "~/Repos/ligrec_decoupleR/input/test_em.csv",
  ann_path = "~/Repos/ligrec_decoupleR/input/test_metadata.csv",
  output_path = "~/Repos/ligrec_decoupleR/output/NATMI_test",
  .assay = "SCT",
  .format = TRUE,
  .write_data = FALSE,
  .subsampling_pipe = FALSE,
  .seed = 1004
)
}
\arguments{
\item{omni_resources}{List of OmniPath resources}

\item{omnidbs_path}{path of saved omnipath resources}

\item{natmi_path}{path of NATMI code and dbs}

\item{em_path}{expression matrix path}

\item{ann_path}{annotations (i.e. clusters) path}

\item{output_path}{NATMI output path}

\item{.format}{bool whether to format output}

\item{.write_data}{bool whether Extract data from Seurat Object}

\item{.subsampling_pipe}{bool whether ran as part of the robustness pipe:
if true, we use Seurat object name to modify output and input, so that
each subsampling is saved to a different dir}

\item{.default_run}{bool whether to run default DBs or not}
}
\value{
DF with NATMI results
}
\description{
Call NATMI Pipeline from R with OmniPath
}
\details{
This function will take omnipath resources saved to csvs and copy
 them to the NATMI dbs folder, then it will natively call the python modules
 in the NATMI dir and save the output into a specified directory.
 It will then load and format the output to a DF.

NB! Please stick to full paths.

NATMI Arguments:
  --interDB INTERDB
                        lrc2p (default) has literature supported ligand-receptor pairs | lrc2a has putative and literature supported ligand-receptor pairs | the user-supplied interaction database can also be used by calling the name of database file without extension
  --interSpecies INTERSPECIES
                        human (default) | mouse | expandp | expanda
  --emFile EMFILE       the path to the normalised expression matrix file with row names (gene identifiers) and column names (cell-type/single-cell identifiers)
  --annFile ANNFILE     the path to the metafile in which column one has single-cell identifiers and column two has corresponding cluster IDs (see file 'toy.sc.ann.txt' as an example)
  --species SPECIES     human (default) | mouse | rat | zebrafish | fruitfly | chimpanzee | dog | monkey | cattle | chicken | frog | mosquito | nematode | thalecress | rice | riceblastfungus | bakeryeast | neurosporacrassa | fissionyeast | eremotheciumgossypii | kluyveromyceslactis, 21 species are supported
  --idType IDTYPE       symbol (default) | entrez(https://www.ncbi.nlm.nih.gov/gene) | ensembl(https://www.ensembl.org/) | uniprot(https://www.uniprot.org/) | hgnc(https://www.genenames.org/) | mgi(http://www.informatics.jax.org/mgihome/nomen/index.shtml) | custom(gene identifier used in the expression matrix)
  --coreNum CORENUM     the number of CPU cores used, default is one
  --out OUT             the path to save the analysis results

Stats:
1) The mean-expression edge weights are calculated by multiplying the mean-expression level of the ligand in the
  sending cell type by the mean expression of the receptor in the target cell type (no discriminatory information)
2) The specificity-based edge weights, help identify the most specific edges in the network
 where each specificity is defined as the mean expression of the ligand/receptor in a given cell type
 divided by the sum of the mean expression of that ligand/receptor across all cell types
 # * a weight of 1 means both the ligand and receptor are only expressed in one (not necessarily the same) cell type
 @import rprojroot
}
