% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sca_pipe.R
\name{call_sca}
\alias{call_sca}
\title{Function to call SingleCellSignalR with databases from OmniPath}
\usage{
call_sca(op_resource, seurat_object, .format = TRUE, assay = "SCT", ...)
}
\arguments{
\item{op_resource}{OmniPath Intercell Resource DN}

\item{seurat_object}{Seurat object as input}

\item{.format}{bool whether to format output}

\item{assay}{Seurat assay data to use}
}
\value{
An unfiltered iTALK df sorted by relevance
}
\description{
Function to call SingleCellSignalR with databases from OmniPath
}
\details{
Stats:
LRScore = sqrt(LR product)/mean(raw counts) * sqrt(LR product) where
expression of l > 0 and r > 0
LRScore = 1 is the highest (~ most likely hit), 0 is the lowest.
}
