% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squidpy_pipe.R
\name{call_squidpy}
\alias{call_squidpy}
\title{Call Squidpy Pipeline via reticulate with OmniPath and format results [[DEPRECATED]]}
\usage{
call_squidpy(
  sce,
  op_resource,
  seed = 1004,
  conda_env = NULL,
  assay = "RNA",
  assay.type = "logcounts",
  ...
)
}
\arguments{
\item{sce}{SingleCellExperiment or Seurat Object as input}

\item{op_resource}{Tibble or list of OmniPath resources, typically obtained via
\code{\link{select_resource}}}

\item{seed}{seed passed to squidpy's ligrec function}

\item{conda_env}{python conda environment to run Squidpy; set to liana_env by default}

\item{assay}{assay name}

\item{...}{kwargs passed to Squidpy; For more information see:
\link{https://squidpy.readthedocs.io/en/latest/api/squidpy.gr.ligrec.html#squidpy.gr.ligrec}}

\item{expr_prop}{minimum proportion of gene expression per cell type (0 by default),
yet perhaps one should consider setting this to an appropriate value between 0 and 1,
as an assumptions of these method is that communication is coordinated at the cluster level.}
}
\value{
A list of Squidpy results for each resource
}
\description{
Call Squidpy Pipeline via reticulate with OmniPath and format results [[DEPRECATED]]
}
\details{
CellPhoneDBv2 algorithm re-implementation in Python.

Note that `cluster_key` is a parameter passed to the Squidpy function,
by default this will be set to the default Ident of the Seurat object.
}
