% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squidpy_pipe.R
\name{call_squidpyR}
\alias{call_squidpyR}
\title{Call Squidpy Pipeline via reticulate with OmniPath and format results}
\usage{
call_squidpyR(seurat_object, omni_resources, python_path, .seed = 1004, ...)
}
\arguments{
\item{seurat_object}{Seurat object as input}

\item{omni_resources}{List of OmniPath resources}

\item{python_path}{path to python version to use in reticulate}

\item{.seed}{used to python seed}
}
\value{
A list of Squidpy results for each resource
}
\description{
Call Squidpy Pipeline via reticulate with OmniPath and format results
}
\details{
CellPhoneDB v2 algorithm implementation in Python
Stats:
Mean expr
pval from shuffled cluster
}
