% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellchat_pipe.R
\name{cellchat_formatDB}
\alias{cellchat_formatDB}
\title{Helper Function to Format CellChatDB}
\usage{
cellchat_formatDB(op_resource, exclude_anns)
}
\arguments{
\item{op_resource}{OmniPath Intercell Resource DN}

\item{exclude_anns}{Annotation criteria to be excluded}
}
\description{
Helper Function to Format CellChatDB
}
