% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_omnipath.R
\name{check_if_dissociated}
\alias{check_if_dissociated}
\title{Helper Function to check if there are dissociated entities which also exist as
complexes.}
\usage{
check_if_dissociated(complex_omni, check_entity, anchor_entity)
}
\arguments{
\item{complex_omni}{an OmniPath resource with complexes}

\item{check_entity}{the entity to be check for duplicates}

\item{anchor_entity}{the anchor entity with which we check for duplicates}
}
\description{
Helper Function to check if there are dissociated entities which also exist as
complexes.
}
\details{
We count the times that a ligand (check_entity) exists in a combination
with the same receptor (anchor_entity)
}
