% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_plot.R
\name{chord_freq}
\alias{chord_freq}
\title{Frequency ChordDiagram}
\usage{
chord_freq(
  liana_res,
  source_groups = NULL,
  target_groups = NULL,
  cex = 1,
  transparency = 0.4,
  facing = "clockwise",
  adj = c(-0.5, 0.05),
  ...
)
}
\arguments{
\item{liana_res}{aggregated `liana_wrap` results from multiple methods,
or alternatively results from running `liana_wrap` with a single method.
Should be filtered by some condition (e.g. preferential consesus ranking,
specific interactions, etc).}

\item{source_groups}{names of the source (sender) cell types (NULL = no filter)}

\item{target_groups}{names of the target cell types (NULL = no filter)}

\item{cex}{label relative font size}

\item{transparency}{transparency}

\item{facing}{axis label rotation (check `circlize::circos.text` for options)}

\item{...}{other paramters passed to `circlize::chordDiagram`}

\item{offset}{for text.}
}
\description{
Frequency ChordDiagram
}
