% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_descriptive.R
\name{classes_bar}
\alias{classes_bar}
\title{Stacked barplot from a data frame of classified entities}
\usage{
classes_bar(data, entity, resource, var)
}
\arguments{
\item{data}{A data frame with classified entities (ligands, receptors or
interactions).}

\item{entity}{The name of the entity, to be included in the output file
name and the y axis label.}

\item{resource}{The name of the resource, to be included in the output
file name.}

\item{var}{Name of the classifying variable.}
}
\value{
Returns `NULL`.
}
\description{
Stacked barplot from a data frame of classified entities
}
