% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_descriptive.R
\name{cluster_for_heatmap}
\alias{cluster_for_heatmap}
\title{Rearranges a data frame before plotting a heatmap}
\usage{
cluster_for_heatmap(data, cat1, cat2, val)
}
\arguments{
\item{data}{A data frame used for the heatmap.}

\item{cat1}{Name of the first categorical variable (one axis of the
heatmap).}

\item{cat2}{Name of the second categorical variable (another axis of the
heatmap).}

\item{val}{Name of the numeric variable corresponding to the color scale
of the heatmap.}
}
\description{
Applies hierarchical clustering on two variables and converts them to
factors with the order of their levels corresponding to the clusters.
}
