% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_omnipath.R
\name{compile_ligrec}
\alias{compile_ligrec}
\title{Function to get unfiltered intercell resources}
\usage{
compile_ligrec(omni_variants = FALSE, lr_pipeline = TRUE)
}
\arguments{
\item{omni_variants}{bool whether to get different OmniPath variants (e.g.
_full, based on ligrec resource quality quartile, or if only lig_rec)}

\item{lr_pipeline}{bool whether to format for lr_pipeline and remove
duplicate LRs (mainly from composite OmniDB due to category
(adhesion vs lr))}
}
\description{
For each resource and OmniPath variant compiles tables of ligands,
receptors and interactions.
}
\details{
calls on \code{omnipath_intercell}, \code{intercell_connections},
    \code{get_partners}, and \code{intercell_connections}
}
