% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_omnipath.R
\name{compile_ligrec_descr}
\alias{compile_ligrec_descr}
\title{Ligand-receptor data for the descriptive part}
\usage{
compile_ligrec_descr()
}
\description{
Ligands, receptors and connections from each resource in a nested list
of tibbles.
}
\seealso{
\code{\link{compile_ligrec}}
}
