% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytotalk.R
\name{compute_mi_dist}
\alias{compute_mi_dist}
\title{Compute mutual information distance from expression vectors}
\usage{
compute_mi_dist(exp1, exp2, nbins)
}
\arguments{
\item{exp1}{A vector containing the normalized gene expression for the first
gene (ligand)}

\item{exp2}{A vector containing the normalized gene expression for the first
gene (receptor)}

\item{nbins}{Number of bins to discretize the expression values.}
}
\value{
The mututal information distance, also refered to as non-self atalk score
}
\description{
Given two normalized gene expression vectors, and a given number of bins, this
function uses the entropy package to compute the mutual information between the
two vectors. Values passed to this function should be log-transformed.
}
