% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstalk_funs.R
\name{compute_nst_from_matrix}
\alias{compute_nst_from_matrix}
\title{Compute non-self talk scores from matrix}
\usage{
compute_nst_from_matrix(mat, ligand_receptor_df, seed)
}
\arguments{
\item{mat}{A matrix containing the log-transformed normalized expression values.}

\item{ligand_receptor_df}{A data frame with, at least, two columns named 'ligand'
and 'receptor' containing the ligand-receptor pairs to evaluate}
}
\value{
A data frame with the non-self-talk score for each pair of genes (ligand-receptors).
}
\description{
Calculates the non-self talk scores (also refered to as mutual information distances)
for each ligand-receptor pair using the normalized expression matrix for a given
cell type. It expects log-transformed expression values.
}
