% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytotalk_funs.R
\name{compute_nst_scores}
\alias{compute_nst_scores}
\title{Compute non-self talk scores from SingleCellExperiment object}
\usage{
compute_nst_scores(
  sce,
  ligand_receptor_df,
  assay.type = "logcounts",
  seed = 1004
)
}
\arguments{
\item{sce}{A SingleCellExperiment object containing log-normalized expression
values and cell type annotations as colLabels}

\item{ligand_receptor_df}{A data frame with, at least, two columns named 'ligand'
and 'receptor' containing the ligand-receptor pairs to evaluate}

\item{assay.type}{The name of the data slot containing the log-normalized expression
values in the SingleCellExperiment object}
}
\value{
A data frame with the computed non-self-talk score for each ligand-receptor
pair on each cell-type.
}
\description{
This function computes the non-self talk scores for all the cell types contained
in a SingleCellExperiment object. It iterates through the ligand-receptor pairs
that are provided as input and calculates these entropy-based measurements for each
cell type using its normalized expression matrix, which should contain log-transformed
values.
}
