% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytotalk.R
\name{compute_pem_scores}
\alias{compute_pem_scores}
\title{Compute Preferential Expression Measure scores from SingleCellExperiment object}
\usage{
compute_pem_scores(sce, assay.type = "logcounts")
}
\arguments{
\item{sce}{A SingleCellExperiment object containing log-normalized expression
values and cell type annotations as colLabels}

\item{assay.type}{The name of the data slot containing the log-normalized expression
values in the SingleCellExperiment object}
}
\value{
A matrix containing the computed PEM values
}
\description{
Uses the information contained in a SingleCellExperiment object to stratify the
expression matrix by cell type, and then computes the PEM scores from the exponential
of the normalized expression values. IMPORTANT: This function expects an object
containing log-transformed values (not raw counts).
}
