% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectome_pipe.R
\name{conn_formatDB}
\alias{conn_formatDB}
\title{Helper Function to convert Omni to Connectome resource Format}
\usage{
conn_formatDB(op_resource)
}
\arguments{
\item{op_resource}{OmniPath resource}
}
\description{
Helper Function to convert Omni to Connectome resource Format
}
\keyword{internal}
