% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_call.R
\name{cpdb_score}
\alias{cpdb_score}
\title{Function to calculate pvalues as in CPDB}
\usage{
cpdb_score(
  lr_res,
  sce_mat,
  nperms = 100,
  seed = 1234,
  trim = 0.1,
  parallelize = TRUE,
  workers = 4
)
}
\arguments{
\item{sce_mat}{single cell expression matrix (transposed)}
}
\value{
lr_res with pvalues
}
\description{
Function to calculate pvalues as in CPDB
}
