% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstalk_funs.R
\name{cytotalk_score}
\alias{cytotalk_score}
\title{Compute cross-talk score from a Seurat Object}
\usage{
cytotalk_score(lr_res, sce, score_col, assay.type, seed, ...)
}
\description{
This function executes all the required functions to extract and compute the
cross-talk scores as defined by the Cytotalk authors. .
}
