% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{descriptive_plots}
\alias{descriptive_plots}
\title{Ligand-receptor resources descriptive plots}
\usage{
descriptive_plots(ligrec = NULL, outdir = NULL)
}
\description{
Generates a number of comparative and descriptive plots about
ligand-receptor resources.
}
