% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_aggregate.R
\name{.aggregate_rank}
\alias{.aggregate_rank}
\title{Robust Aggregate ranks using `RobustRankAggreg`}
\usage{
.aggregate_rank(liana_mlist, join_cols, ...)
}
\arguments{
\item{liana_mlist}{liana list with method tibbles}

\item{join_cols}{columns to be concatenated to create entity to be ranked}

\item{...}{
  Arguments passed on to \code{\link[RobustRankAggreg:aggregateRanks]{RobustRankAggreg::aggregateRanks}}
  \describe{
    \item{\code{glist}}{list of element vectors, the order of the
  vectors is used as the ranking.}
    \item{\code{rmat}}{the rankings in matrix format. The glist is
  by default converted to this format.}
    \item{\code{N}}{the number of ranked elements, important when
  using only top-k ranks, by default it is calculated as
  the number of unique elements in the input.}
    \item{\code{method}}{rank aggregation method, by defaylt
  \code{'RRA'}, other options are \code{'min'},
  \code{'geom.mean'}, \code{'mean'}, \code{'median'} and
  \code{'stuart'}}
    \item{\code{full}}{indicates if the full rankings are given,
  used if the the sets of ranked elements do not match
  perfectly}
    \item{\code{exact}}{indicator showing if exact p-value will be
  calculated based on rho score (Default: if number of
  lists smaller than 10, exact is used)}
    \item{\code{topCutoff}}{a vector of cutoff values used to limit
  the number of elements in the input lists elements do not
  match perfectly}
  }}
}
\description{
Robust Aggregate ranks using `RobustRankAggreg`
}
