% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_pipe.R
\name{.antilog1m}
\alias{.antilog1m}
\title{Helper function to inverse logged counts}
\usage{
.antilog1m(x, base = 2)
}
\arguments{
\item{x}{mat or array}

\item{base}{base for conversion}
}
\description{
Helper function to inverse logged counts
}
