% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_prep.R
\name{.filter_sce}
\alias{.filter_sce}
\title{Helper function to perform basic filterin on the SCE object prior to feeding it to LIANA}
\usage{
.filter_sce(sce, verbose)
}
\arguments{
\item{sce}{SingleCellExperiment Object}
}
\value{
SingleCellExperiment object
}
\description{
Helper function to perform basic filterin on the SCE object prior to feeding it to LIANA
}
