% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_prep.R
\name{.format_idents}
\alias{.format_idents}
\title{Helper Function to get/format the required indentity if required}
\usage{
.format_idents(metadata, active_idents, idents_col, object_class, verbose)
}
\arguments{
\item{metadata}{metadata obtained from the sce object}

\item{active_idents}{active idents (e.g. Seurat::Idents, or sce colLabels)}

\item{idents_col}{the cell identities/labels to be used. By default set to NULL, and will used the active
Idents or colLabels from the seurat_object or SCE, respectively.}

\item{object_class}{class of the sce object (SingleCellExperiment or Seurat)}

\item{verbose}{logical whether to output messages and warnings (`TRUE` by default)}
}
\description{
Helper Function to get/format the required indentity if required
}
