% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_utils.R
\name{.generate_complex_dict}
\alias{.generate_complex_dict}
\title{Helper function to generate a dictionary also for the complexes}
\usage{
.generate_complex_dict(translated_subunits, entity = "target")
}
\arguments{
\item{translated_subunits}{decomplexified op_resource with already recoded
subunits}

\item{entity}{column of interest (target or source)}
}
\description{
Helper function to generate a dictionary also for the complexes
}
