% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_plot.R
\name{.get_freq}
\alias{.get_freq}
\title{Helper function to obtain interaction frequencies}
\usage{
.get_freq(liana_res)
}
\arguments{
\item{liana_res}{liana-formatted results}
}
\description{
Helper function to obtain interaction frequencies
}
