% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_utils.R
\name{.join_meta}
\alias{.join_meta}
\title{Join sample descriptor column from SCE to another table}
\usage{
.join_meta(sce, right, sample_col, group_col)
}
\arguments{
\item{sce}{SingleCellExperiment}

\item{right}{(df) to join to colData per sample}

\item{sample_col}{unique identifier column from colData}

\item{group_col}{sample_col descriptor column}
}
\description{
Join sample descriptor column from SCE to another table
}
