% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_aggregate.R
\name{.liana_consensus}
\alias{.liana_consensus}
\title{Get Consensus Rankings for the Methods}
\usage{
.liana_consensus(liana_aggr, cap)
}
\arguments{
\item{liana_aggr}{Aggregated method results}

\item{cap}{Value assigned to NA (by default, the max number of all possible
interactions, depending on the resource)}
}
\value{
aggregated liana tibble with consensus ranks
}
\description{
Get Consensus Rankings for the Methods
}
