% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_wrapper.R
\name{.list2tib}
\alias{.list2tib}
\title{Helper Function to  Handle list or not list method calls}
\usage{
.list2tib(res)
}
\arguments{
\item{res}{list of lists (e.g. Method-Resource Results from the LIANA pipe)}
}
\value{
The first element of the list
}
\description{
Helper Function to  Handle list or not list method calls
}
