% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_pipe.R
\name{.prep_universe}
\alias{.prep_universe}
\title{Helper function to format SCE to the LR universe}
\usage{
.prep_universe(sce, entity_genes, verbose)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{entity_genes}{union of ligand-receptor genes}

\item{verbose}{verbose - True/Flase}
}
\description{
Helper function to format SCE to the LR universe
}
