% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_aggregate.R
\name{.rank_matrix}
\alias{.rank_matrix}
\title{Function to convert a list of characters to a ranked matrix [0,1]}
\usage{
.rank_matrix(glist, ...)
}
\arguments{
\item{glist}{a list of ranked/ordered characters}
}
\value{
a matrix filled with 0-1 values depending on the position order
of the characters.
}
\description{
Function to convert a list of characters to a ranked matrix [0,1]
}
\details{
Adated from Kolde et al., 2012.
Required due to the removal of the RobustRankAggregate package from CRAN.
}
\references{
Kolde, R., Laur, S., Adler, P. and Vilo, J., 2012.
 Robust rank aggregation for gene list integration and meta-analysis.
 Bioinformatics, 28(4), pp.573-580.
}
\keyword{internal}
