% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_specs.R
\name{.score_housekeep}
\alias{.score_housekeep}
\title{Helper function to call aggregate housekeeping scores of external methods.}
\usage{
.score_housekeep()
}
\description{
Helper function to call aggregate housekeeping scores of external methods.
}
\details{
functions the same way as .score_specs, but is only used in
liana_aggragate for the purpose of the manuscript.
}
