% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_aggregate.R
\name{.select_cap}
\alias{.select_cap}
\title{Helper function to execute a function on the vector representing the number
 of rows in the results for each method}
\usage{
.select_cap(liana_res, fun)
}
\arguments{
\item{liana_res}{ligand-receptor stats between clusters, output of
`liana_pipe`}

\item{fun}{function to execute}
}
\description{
Helper function to execute a function on the vector representing the number
 of rows in the results for each method
}
