% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_wrapper.R
\name{.select_method}
\alias{.select_method}
\title{Function to return the appropriate method(s) to be executed}
\usage{
.select_method(method)
}
\arguments{
\item{method}{name of the method}
}
\value{
A list of method function names (to be called by the LIANA wrapper)
}
\description{
Function to return the appropriate method(s) to be executed
}
\details{
Adapted from (jvelezmagic); decoupleR\https://github.com/saezlab/decoupleR/
}
