% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_descriptive.R
\name{enrich2}
\alias{enrich2}
\title{Pairwise enrichment between two factors}
\usage{
enrich2(
  data,
  var1,
  var2 = resource,
  p_adj_method = "fdr",
  test_method = "fisher",
  ...
)
}
\arguments{
\item{data}{A data frame with entities labeled with two categorical
variables \code{var1} and \code{var2}.}

\item{var1}{Name of the first categorical variable.}

\item{var2}{Name of the second categorical variable. Because in
ligand-receptor data it's always `resource` the default is `resource`.}

\item{p_adj_method}{Adjustment method for multiple testing p-value
correction (see \code{stats::p.adjust}).}

\item{test_method}{Characted: either "barnard", "barnard2" or "fisher".
"barnard" uses \code{Barnard::barnard.test} while "barnard2" uses
\code{DescTools::BarnardTest}.}

\item{...}{Passed to the function executing the Barnard test.}
}
\value{
A data frame with all possible combinations of the two categorical
    variables, p-values, adjusted p-values and odds ratios.
}
\description{
From a data frame with two categorical variables performs Barnard or
Fisher tests between all combinations of the levels of the two variables.
In each test the contingency table looks like: in \code{var1} belongs to
level x or not vs. in \code{var2} belongs to level y or not. By default
Fisher test is used simply because Barnard tests take very very long to
compute.
}
