% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_pipe.R
\name{fast_mean}
\alias{fast_mean}
\title{Helper function to fix r mean}
\usage{
fast_mean(mat)
}
\arguments{
\item{mat}{a matrix}
}
\description{
Helper function to fix r mean
}
\details{
r mean is slow and it overflows on memory.
}
