% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_tensor.R
\name{format_c2c_factors}
\alias{format_c2c_factors}
\title{Helper function to format factors from c2c output}
\usage{
format_c2c_factors(
  factors,
  contexts = "contexts",
  interactions = "interactions",
  senders = "senders",
  receivers = "receivers",
  key_sep = "[|]"
)
}
\arguments{
\item{factors}{factors returned from the
`cell2cell.tensor.compute_tensor_factorization` function}

\item{contexts}{name for `contexts` dimension}

\item{interactions}{name for `interactions` dimension}

\item{senders}{name for `senders` dimension}

\item{receivers}{name for `receivers` dimension}

\item{key_sep}{separator for the sample and condition names. These are
used to jointly name the contexts - technically the names of the
`context_df_dict` list passed to the `liana_tensor_c2c` function.}
}
\description{
Helper function to format factors from c2c output
}
\details{
The parameters here would correspond to `order_labels`
}
\keyword{internal}
