% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_crc_data.R
\name{format_crc_meta}
\alias{format_crc_meta}
\title{Helper function to format metadata for CRC data sets}
\usage{
format_crc_meta(crc_seurat)
}
\arguments{
\item{crc_seurat}{CRC Seurat Object}
}
\value{
Return a formatted Metadata for CCC inference
}
\description{
Helper function to format metadata for CRC data sets
}
