% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{format_rank_frequencies}
\alias{format_rank_frequencies}
\title{Helper function to get cell pair activity from rank averages}
\usage{
format_rank_frequencies(result, score_col, .desc_order = TRUE)
}
\arguments{
\item{result}{Result for a specific Tool-resource combinations}

\item{score_col}{Score column name provided by the tool}

\item{.desc_order}{Whether the most significant hits are in descending order,
i.e. the highest is the most sig}
}
\description{
Helper function to get cell pair activity from rank averages
}
\details{
Cell pair ranks are averaged, then converted to z-scores, which
are multiplied by -1 as we want the lowest average ranks to have the highest
z scores
}
