% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectome_pipe.R
\name{FormatConnectome}
\alias{FormatConnectome}
\title{Helper function to filter and format connectome}
\usage{
FormatConnectome(conn, ...)
}
\arguments{
\item{conn}{connectome object}

\item{.log}{whether to log weight}
}
\description{
Helper function to filter and format connectome
}
