% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iTALK_pipe.R
\name{FormatiTALK}
\alias{FormatiTALK}
\title{Helper Function to Filter and format iTalk results}
\usage{
FormatiTALK(italk_res, remove.na = TRUE, .DE = FALSE)
}
\arguments{
\item{italk_res}{iTalk results object}

\item{remove.na}{bool whether to filter NA}
}
\description{
Helper Function to Filter and format iTalk results
}
