% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NATMI_pipe.R
\name{FormatNatmi}
\alias{FormatNatmi}
\title{Load NATMI results from folder and format appropriately}
\usage{
FormatNatmi(output_path, resource_names, .format = TRUE)
}
\arguments{
\item{output_path}{NATMI output path}

\item{resource_names}{results for which resources to load}

\item{.format}{bool whether to format output}
}
\value{
A list of NATMI results per resource loaded from the output
    directory.
}
\description{
Load NATMI results from folder and format appropriately
}
