% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sca_pipe.R
\name{FormatSCA}
\alias{FormatSCA}
\title{Helper function to format SingleCellSignalR results}
\usage{
FormatSCA(sca_res, remove.na = TRUE)
}
\arguments{
\item{sca_res}{Unformatted SCA results}

\item{remove.na}{bool whether to filter SCA output}
}
\description{
Helper function to format SingleCellSignalR results
}
