% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_omnipath.R
\name{geneid_to_uniprot}
\alias{geneid_to_uniprot}
\title{Helper Function to translate to UniProt}
\usage{
geneid_to_uniprot(st, dict)
}
\arguments{
\item{st}{any genesymbol string - to be separate by `_`}

\item{dict}{dfgdfg}
}
\description{
Helper Function to translate to UniProt
}
