% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_ortho.R
\name{generate_homologs}
\alias{generate_homologs}
\title{Function to generate a homologous OmniPath resource}
\usage{
generate_homologs(
  op_resource,
  target_organism,
  max_homologs = 5,
  .missing_fun = NULL,
  symbols_dict = NULL,
  source = "source_genesymbol",
  target = "target_genesymbol",
  verbose = TRUE
)
}
\arguments{
\item{op_resource}{a resource in the format of OmniPath/LIANA}

\item{target_organism}{`ncbi_taxid` or `name` of the target organism.
See `show_homologene` for available organisms via OmnipathR's `HomoloGene`}

\item{max_homologs}{Determines the max number of homologs to be translated.
Certain genes will have multiple homolog matches, with some having also
certain isoforms considered. To exclude cases in which the number of
matched homologs is too high, one can adjust the homologs parameter.
Setting this to `1` would mean that one-to-many homolog matches are discarded}

\item{.missing_fun}{approach to handle missing interactions. By default
set to `NULL` which would mean that any interactions without a homology
match will be filtered. This can be set to e.g. `str_to_title` when working
with murine symbols. Then if a gene has no matched homolog, instead of
discarding it, the `.missing_fun` will be used to format the name from human.
Hence, increasing the number of matches, but likely introducing some
mismatches.}

\item{symbols_dict}{`NULL` by default, then `get_homologene_dict` is called
to generate a dictionary from OmniPathR's homologene resource. Alternatively,
one can pass their own symbols_dictionary.}

\item{source}{name of the source (ligand) column}

\item{target}{name of the target (receptor) column}

\item{verbose}{logical for verbosity}
}
\value{
a converted ligand-receptor resource
}
\description{
Function to generate a homologous OmniPath resource
}
