% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_ortho.R
\name{generate_orthologs}
\alias{generate_orthologs}
\title{Deprecated call to generate_homologs}
\usage{
generate_orthologs(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=generate_orthologs]{generate_orthologs}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
Deprecated call to generate_homologs
}
