% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_omnipath.R
\name{genesymbol_to_uniprot}
\alias{genesymbol_to_uniprot}
\title{Helper Function to translate to UniProt}
\usage{
genesymbol_to_uniprot(st)
}
\arguments{
\item{st}{any genesymbol string - to be separate by `_`}
}
\description{
Helper Function to translate to UniProt
}
\keyword{internal}
