% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_utils.R
\name{get_abundance_summary}
\alias{get_abundance_summary}
\title{Function to get abundance summary}
\usage{
get_abundance_summary(
  sce,
  sample_col,
  idents_col,
  min_cells = 10,
  min_samples = 3,
  min_prop = 0.2
)
}
\arguments{
\item{sce}{SingleCellExperiment Object}

\item{sample_col}{column with sample ids}

\item{idents_col}{column with cell identity (cell type) ids}

\item{min_cells}{minimum cells per identity in each sample}

\item{min_samples}{minimum samples per cell identity}

\item{min_prop}{minimum proportion of samples in which a cell identity is
present (with at least `min_cells`)}
}
\value{
a tibble
}
\description{
Function to get abundance summary
}
\keyword{internal}
