% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{get_bc_heatmap}
\alias{get_bc_heatmap}
\title{Bray-Curtis Heatmap Function}
\usage{
get_bc_heatmap(sig_list, ...)
}
\arguments{
\item{sig_list}{list of significant hits}

\item{...}{
  Arguments passed on to \code{\link[=pheatmap]{pheatmap}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
Bray-Curtis Heatmap Function
}
