% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_utils.R
\name{get_binary_df}
\alias{get_binary_df}
\title{Helper Function to get a binary top hits DF (for all method-resource combos)}
\usage{
get_binary_df(sig_list)
}
\arguments{
\item{sig_list}{list of significant hits per method-resource combo}
}
\description{
Helper Function to get a binary top hits DF (for all method-resource combos)
}
