% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_omnipath.R
\name{get_cellcall}
\alias{get_cellcall}
\title{Function to Obtain the CellCall database}
\usage{
get_cellcall()
}
\value{
cellcall db converted to LIANA/OP format
}
\description{
Function to Obtain the CellCall database
}
