% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_utils.R
\name{get_cellnum}
\alias{get_cellnum}
\title{Helper function to get cell number per cell type}
\usage{
get_cellnum(seurat_path)
}
\arguments{
\item{seurat_path}{Path to Seurat object of interest}
}
\value{
A tibble with Cell_subtype and Cell Number columns
}
\description{
Helper function to get cell number per cell type
}
