% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_call.R
\name{get_connectome}
\alias{get_connectome}
\title{Function to obtain connectome-like weights}
\usage{
get_connectome(seurat_object, op_resource, ...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=.liana_call]{.liana_call}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
Returns a tibble with specificity weights (`weight_sc`) as calculated
   by connectome
}
\description{
Function to obtain connectome-like weights
}
