% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_omnipath.R
\name{get_curated_omni}
\alias{get_curated_omni}
\title{Function to Generate the Curated (Default) LIANA resource}
\usage{
get_curated_omni(
  curated_resources = c("CellPhoneDB", "CellChatDB", "ICELLNET", "connectomeDB2020",
    "CellTalkDB"),
  ...
)
}
\arguments{
\item{curated_resources}{the curated resources from which we wish to obtain interactions.
By default, it includes interactions curated in the context of CCC from CellPhoneDB,
CellChat, ICELLNET, connectomeDB, CellTalkDB, and SignaLink.}
}
\value{
a curated OmniPath resource formatted for LIANA
}
\description{
Function to Generate the Curated (Default) LIANA resource
}
\details{
Here, we define the curated resources as those that are defined as manually
or expert curated in the context of cell-cell communication.
Albeit, "Guide2Pharma", "HPMR", and "Kirouac2010" are also such resources the remainder
of the resources used to generate Omnipath, use those as sources.
Hence, we assume that the second round of manual curation done in subsequent, more recently published
resources would already contain the high quality interactions of the aforementioned 3.
We also omit Cellinker, as it results in a large mount of ambigous interactions, but
one could consider adding it to the list of curated resources.
}
