% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_pipe.R
\name{get_global_mean}
\alias{get_global_mean}
\title{Helper function to obtain a global mean}
\usage{
get_global_mean(seurat_object, assay, assay.type)
}
\arguments{
\item{seurat_object}{Seurat object}

\item{assay}{assay to be used ("RNA" by default)}

\item{assay.type}{- the type of data to be used to calculate the means
(counts by default), available options are: "counts" and "logcounts"}
}
\value{
a single value for the whole dataset
}
\description{
Helper function to obtain a global mean
}
\details{
global mean is used in the `SingleCellSignalR` scores.
}
