% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_ortho.R
\name{get_homologene_dict}
\alias{get_homologene_dict}
\title{Helper function to get homologene dictionary}
\usage{
get_homologene_dict(entities, target_organism)
}
\arguments{
\item{entities}{genes to be converted}

\item{target_organism}{target organism (obtain tax id from `show_homologene`)}
}
\description{
Helper function to get homologene dictionary
}
\keyword{internal}
