% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_descriptive.R
\name{get_intersect}
\alias{get_intersect}
\title{Function to get the intersect of vectors within the same list}
\usage{
get_intersect(vector_list, .names)
}
\arguments{
\item{vector_list}{List of character vectors (i.e. interactions per resource)}

\item{.names}{names of the list elements}
}
\description{
Function to get the intersect of vectors within the same list
}
