% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_utils.R
\name{get_jacc}
\alias{get_jacc}
\title{Get Jaccard indeces between specific combination of resources or methods}
\usage{
get_jacc(sig_list, methods, resources)
}
\arguments{
\item{sig_list}{List of significant hits per Method-resource combination.
Named list of methods with each element being a named list of resources.}

\item{methods}{character vector of method names}

\item{resources}{character vector of resource names}
}
\value{
Returns a Jaccard index between all combinations of the supplied
  resources and methods
}
\description{
Get Jaccard indeces between specific combination of resources or methods
}
