% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_omnipath.R
\name{get_ligands}
\alias{get_ligands}
\title{Retrieves ligands from one ligand receptor resource}
\usage{
get_ligands(resource, ...)
}
\arguments{
\item{resource}{Name of the resource. For a full list of resources
see \code{get_lr_resources}.}

\item{...}{
  Arguments passed on to \code{\link[=intercell_connections]{intercell_connections}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
Retrieves ligands from one ligand receptor resource
}
