% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_call.R
\name{get_logfc}
\alias{get_logfc}
\title{Function to obtain logFC weights}
\usage{
get_logfc(seurat_object, op_resource, ...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=.liana_call]{.liana_call}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
Returns a tibble with a logFC metric (`logfc_comb`). `logfc_comb` is
   calculated as the product of the (1 vs the rest) log2FC for each ligand
   and receptor gene
}
\description{
Function to obtain logFC weights
}
