% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_omnipath.R
\name{get_lr_resources}
\alias{get_lr_resources}
\title{Helper function that returns the name of each intercell resource in OmniPath}
\usage{
get_lr_resources()
}
\value{
A list of strings for each intercell resource in OmniPath
}
\description{
Helper function that returns the name of each intercell resource in OmniPath
}
