% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_call.R
\name{get_natmi}
\alias{get_natmi}
\title{Function to obtain NATMI-like weights}
\usage{
get_natmi(lr_res, ...)
}
\arguments{
\item{lr_res}{ligand-receptor DE results and other stats between clusters}

\item{...}{
  Arguments passed on to \code{\link[=liana_call]{liana_call}}
  \describe{
    \item{\code{method}}{name of the method to be called}
    \item{\code{op_resource}}{resource tibble obtained via \link{liana::select_resource}}
  }}
}
\value{
Returns a tibble with specificity weights (`edge_specificity`)
   as calculated by NATMI
}
\description{
Function to obtain NATMI-like weights
}
