% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_omnipath.R
\name{get_partners}
\alias{get_partners}
\title{Retrieves intercellular communication partners (ligands or receptors) from
one ligand-receptor resource.}
\usage{
get_partners(side, resource, ...)
}
\description{
Retrieves intercellular communication partners (ligands or receptors) from
one ligand-receptor resource.
}
