% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm_funs.R
\name{get_permutations}
\alias{get_permutations}
\title{Helper Function to generate shuffled means}
\usage{
get_permutations(
  lr_res,
  sce,
  nperms = 1000,
  seed = 1234,
  parallelize = FALSE,
  workers = 4,
  assay.type = "logcounts"
)
}
\arguments{
\item{lr_res}{liana_pipe results}

\item{sce}{SingleCellExperiment Object}

\item{nperms}{number of permutations}

\item{seed}{number used to set random seed}

\item{parallelize}{whether to parallelize}

\item{workers}{Number of workers to be used in parallelization}

\item{assay.type}{- the type of data to be used to calculate the means
(counts by default), available options are: "counts" and "logcounts"}
}
\value{
Returns a list of shuffled gene means by cluster
}
\description{
Helper Function to generate shuffled means
}
\details{
This function could be made generalizable to any set of genes,
  depending on the set (currently lr_res genes) that is used to filter - i.e.
  it could be replaced with e.g. genes from TF regulons
}
