% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{get_rank_frequencies}
\alias{get_rank_frequencies}
\title{Helper function to convert list with all resources ranked to frequencies df}
\usage{
get_rank_frequencies(full_list)
}
\arguments{
\item{full_list}{list with all resources ranked to frequencies df}
}
\description{
Helper function to convert list with all resources ranked to frequencies df
}
