% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_omnipath.R
\name{get_receptors}
\alias{get_receptors}
\title{Retrieves receptors from one ligand-receptor resource}
\usage{
get_receptors(resource, ...)
}
\description{
Retrieves receptors from one ligand-receptor resource
}
