% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_call.R
\name{get_sca}
\alias{get_sca}
\title{Function to obtain SingleCellSignalR-like scores}
\usage{
get_sca(lr_res, ...)
}
\arguments{
\item{lr_res}{ligand-receptor DE results and other stats between clusters}

\item{...}{
  Arguments passed on to \code{\link[=liana_call]{liana_call}}
  \describe{
    \item{\code{op_resource}}{resource tibble obtained via \link{liana::select_resource}}
  }}
}
\value{
Returns a tibble with specificity weights (`LRscore`) as calculated
   by SingleCellSignalR
}
\description{
Function to obtain SingleCellSignalR-like scores
}
