% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{get_simdist_heatmap}
\alias{get_simdist_heatmap}
\title{Jaccard Similarities Heatmap Function}
\usage{
get_simdist_heatmap(sig_list, ...)
}
\arguments{
\item{sig_list}{list of significant hits}

\item{...}{
  Arguments passed on to \code{\link[=get_simil_dist]{get_simil_dist}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
Jaccard Similarities Heatmap Function
}
