% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_utils.R
\name{get_simil_dist}
\alias{get_simil_dist}
\title{Helper Function to get Similarties and Distances from binary dfs}
\usage{
get_simil_dist(sim_dist = "simil", ...)
}
\value{
a similarity/dissimilarity matrix
}
\description{
Helper Function to get Similarties and Distances from binary dfs
}
