% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{get_swapped_list}
\alias{get_swapped_list}
\title{Helper function to Swap Nested Lists}
\usage{
get_swapped_list(sig_list)
}
\arguments{
\item{sig_list}{named list of significant hits. Named list of methods with
each element being a named list of resources}
}
\value{
A list of resources with each element being a named list of methods
}
\description{
Helper function to Swap Nested Lists
}
\details{
Swap from nested resource lists to nested method lists,
 previously the resources were nested by method, this returns the opposite
}
