% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_utils.R
\name{get_top_hits}
\alias{get_top_hits}
\title{Get top hits}
\usage{
get_top_hits(spec_list, n_ints = c(100, 500, 5000))
}
\arguments{
\item{spec_list}{list of spec objects with ligrec results}
}
\value{
A list of top hits per tool/tool_parameter
}
\description{
Get top hits
}
