% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_omnipath.R
\name{get_up_dict}
\alias{get_up_dict}
\title{Helper function to get UniProt dictionary}
\usage{
get_up_dict(
  ligrec_res,
  up,
  key_column1 = "Ligand_ID",
  key_column2 = "Receptor_ID"
)
}
\arguments{
\item{ligrec_res}{ligand_receptor resource to translate}

\item{up}{uniprot db to be queried}

\item{key_column1}{name of the ligand  column}

\item{key_column2}{name of the receptor column}
}
\description{
Helper function to get UniProt dictionary
}
\keyword{internal}
