% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_pipe.R
\name{\%/*/\%}
\alias{\%/*/\%}
\title{Helper function to multiply or divide depending on sign}
\usage{
x \%/*/\% y
}
\arguments{
\item{x}{target for weighing}

\item{y}{weight}
}
\description{
Helper function to multiply or divide depending on sign
}
