% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_plot.R
\name{heat_freq}
\alias{heat_freq}
\title{Communication Frequency heatmap plot}
\usage{
heat_freq(liana_res, ...)
}
\arguments{
\item{liana_res}{aggregated liana results (preferably truncated
to some threshold)}

\item{...}{
  Arguments passed on to \code{\link[=liana_heatmap]{liana_heatmap}}
  \describe{
    \item{\code{mat}}{Diagonal celltype-celltype matrix to be plotted. In theory,
any metric deemed meaningful between cell pairs can be plotted.}
    \item{\code{font_size}}{base font_size - other fontsizes are relative to this one}
    \item{\code{grid_text}}{logical whether to display grid text or not}
    \item{\code{name}}{name of the heatmap.
By default the heatmap name is used as the title of the heatmap legend.}
    \item{\code{row_title}}{Row tittle}
    \item{\code{column_title}}{Column tittle}
  }}

\item{pallette}{colour palette to be used for the cells}
}
\description{
Communication Frequency heatmap plot
}
\details{
This plot was inspired by CellPhoneDB and also CellChat's heatmap design.
It makes the assumption that the number of interactions inferred between cell
types is informative of the communication events occurring in the system as a whole.
This is a rather strong assumption limited by the arbitrarily filtered
interactions Thus, I suggest that one limits any conclusions, unless supported
by complimentary information, such as biological prior knowledge.
}
