% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_plot.R
\name{heat_spec}
\alias{heat_spec}
\title{Communication Specificity heatmap plot}
\usage{
heat_spec(
  liana_res,
  name = "Average\\nSpecificity",
  pallette = c("white", "#2cb7b9"),
  ...
)
}
\arguments{
\item{liana_res}{aggregated liana results (preferably truncated
to some threshold)}

\item{name}{name of the heatmap.
By default the heatmap name is used as the title of the heatmap legend.}

\item{pallette}{colour palette to be used for the cells}

\item{...}{
  Arguments passed on to \code{\link[=liana_heatmap]{liana_heatmap}}
  \describe{
    \item{\code{mat}}{Diagonal celltype-celltype matrix to be plotted. In theory,
any metric deemed meaningful between cell pairs can be plotted.}
    \item{\code{font_size}}{base font_size - other fontsizes are relative to this one}
    \item{\code{grid_text}}{logical whether to display grid text or not}
    \item{\code{row_title}}{Row tittle}
    \item{\code{column_title}}{Column tittle}
  }}
}
\description{
Communication Specificity heatmap plot
}
\details{
This function builds on the frequency heatmap. Here, we plot the
average specificity per celltype pair, using NATMI's specificity edges.
NATMI's specificity ranges from 1 to 0, with 1 being assigned to interactions
uniquely expressed in a pair of cell types. Similarly to the previous plot,
this one is also limited by the arbitrarity of filtering and any conclusions
should be considered with caution.
}
