% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{hgnc_ligrec_classes}
\alias{hgnc_ligrec_classes}
\title{Ligand-receptor classification from HGNC}
\usage{
hgnc_ligrec_classes(ligrec, largest = 15)
}
\description{
Ligand-receptor classification from HGNC
}
