% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_descriptive.R
\name{interactions_shared}
\alias{interactions_shared}
\title{Overlap between resources pairwise}
\usage{
interactions_shared(ligrec_binary)
}
\arguments{
\item{ligrec_binary}{Binarized data frame with interactions.}
}
\value{
Shared elements between Resource Y and X dataframe
}
\description{
Overlap between resources pairwise
}
