% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_omnipath.R
\name{intercell_connections}
\alias{intercell_connections}
\title{Retrieves the interactions from one ligand-receptor resource}
\usage{
intercell_connections(resource, ...)
}
\arguments{
\item{resource}{Name of the resource. For a full list of resources
see \code{get_lr_resources}.}

\item{...}{Passed to \code{omnipath_intercell} or
\code{OmnipathR::import_post_translational_interactions}.}
}
\description{
Retrieves the interactions from one ligand-receptor resource
}
