% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/italk_pipe.R
\name{italk_formatDB}
\alias{italk_formatDB}
\title{Helper Function to convert Omni to iTALK resource Format}
\usage{
italk_formatDB(op_resource)
}
\arguments{
\item{op_resource}{OmniPath resource}
}
\description{
Helper Function to convert Omni to iTALK resource Format
}
