% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_descriptive.R
\name{jacc_pairwise}
\alias{jacc_pairwise}
\title{Pairwise Jaccard Index between resources}
\usage{
jacc_pairwise(ligrec_binary)
}
\arguments{
\item{ligrec_binary}{}
}
\value{
Long data frame with Jaccard indeces.
}
\description{
Pairwise Jaccard Index between resources
}
