% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_descriptive.R
\name{jaccheat_save}
\alias{jaccheat_save}
\title{Save Jaccard Heatmaps}
\usage{
jaccheat_save(df, plotname, guide_title)
}
\arguments{
\item{df}{Data frame to be pivoted and used to plot the heatmap.}

\item{plotname}{Name of the plot to be saved.}
}
\description{
Save Jaccard Heatmaps
}
