% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_pipe.R
\name{join_log2FC}
\alias{join_log2FC}
\title{Helper Function to join log2FC dataframe to LR_res}
\usage{
join_log2FC(lr_res, logfc_df, source_target, entity)
}
\arguments{
\item{lr_res}{LR formatted DE results from \link{ligrec_degformat}}

\item{logfc_df}{obtained via \link{get_log2FC}}

\item{entity}{ligand or receptor}
}
\description{
Helper Function to join log2FC dataframe to LR_res
}
