% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_descriptive.R
\name{largest_groups}
\alias{largest_groups}
\title{Keep only the largest groups according to a grouping variable}
\usage{
largest_groups(data, var, largest = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{var}{Name of the grouping variable.}

\item{largest}{Numeric: how many of the largest groups to keep. If `NULL`
the data frame will be returned unchanged.}
}
\description{
Keep only the largest groups according to a grouping variable
}
