% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_aggregate.R
\name{liana_aggregate}
\alias{liana_aggregate}
\title{Function to Aggregate CCC Method Results}
\usage{
liana_aggregate(
  liana_res,
  aggregate_how = NULL,
  resource = NULL,
  set_cap = "max",
  cap = NULL,
  get_ranks = TRUE,
  get_agrank = TRUE,
  .score_mode = .score_specs,
  verbose = TRUE,
  join_cols = NULL,
  ...
)
}
\arguments{
\item{liana_res}{LIANA results}

\item{aggregate_how}{way to aggregate, by default (NULL) will aggregate
all passed methods with the approach specified in `liana:::..score_specs`.
Alternative options are `magnitude` and `specificity`.}

\item{resource}{If methods are ran with multiple resources, the name of the
resource of interest needs to be provided
*Note* if a name is not provided, the first results based on the first
resource in the list will be returned}

\item{set_cap}{Function used to set ranked cap (i.e. the value that is
assigned to interactions with NA for scores);
By default, this is set to "max", which is the maximum number of interactions
obtained by the methods; Some methods return all possible ligand-receptor
combinations for each possible source and target cell pair - i.e. the
known universe of all possible interactions (based on the CCC resource)}

\item{cap}{A cap can for all methods can also be manually set, then the top X
interactions, based on the `specificity` scores for each method will be
returned and the ranking will be carried out solely on them}

\item{get_ranks}{boolean, whether to return consensus ranks for methods}

\item{get_agrank}{boolean, whether to return aggregate rank using the
`RobustRankAggreg` package.}

\item{.score_mode}{defines the way that the methods would be aggragate.
By default, we use the score of each method which reflects specificity
(if available), if not e.g. the case of SCA we use it's sole scoring function.
This aggregation is by default done on the basis of the list returns by
`.score_mode`. Alternatively, one could pass `.score_housekeep` to obtain an
aggragate of the housekeeping interactions of each method.}

\item{join_cols}{columns by which different method results will be joined.
NULL by default, and automatically will handle the columns depending on the
methods used.}

\item{...}{
  Arguments passed on to \code{\link[=.rank_matrix]{.rank_matrix}}
  \describe{
    \item{\code{glist}}{a list of ranked/ordered characters}
  }}
}
\value{
Tibble with the interaction results and ranking for each method
}
\description{
Function to Aggregate CCC Method Results
}
\details{
set_cap is the name of the name of a function that is to be executed
  on a vector representing the number of rows in the results for each method,
  by default this is set to \link{base::max}, but any other function that
  works with vectors could be passed - e.g. min, mean, etc.

This function also decomplexifies any complex present in the CellChat results
which returns complexes by default
}
\examples{
liana_path <- system.file(package = "liana")
# load testdata
testdata <- readRDS(file.path(liana_path , "testdata", "input", "testdata.rds"))
# run liana
liana_res <- liana_wrap(testdata, method=c("sca", "natmi"))
# aggregate results from multiple methods
liana_res <- liana_aggregate(liana_res)
}
