% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_mofa.R
\name{liana_cc2mofa}
\alias{liana_cc2mofa}
\title{Wrapper function to run `cell2cell_tensor` with LIANA output.}
\usage{
liana_cc2mofa(
  context_df_dict,
  score_col = "LRscore",
  lr_prop = 0.5,
  lr_delim = "^",
  lr_fill = NA,
  lr_min = 15,
  cell_delim = "&",
  lr_cell_delim = ";",
  key_sep = "|",
  verbose = TRUE
)
}
\arguments{
\item{context_df_dict}{A list of tibbles with liana-style results per contexts
liana format refers to `liana_aggregate` or `liana_wrap` with 1 method output.}

\item{score_col}{name of the score to be used (LRscore, by default)}

\item{lr_prop}{minimum proportion across samples}

\item{lr_delim}{Delimiter that separates ligands and receptors in the
ligand-receptor pair in the edgelists (score matrices)}

\item{lr_fill}{value to fill in missing interactions across contexts
(0 by default)}

\item{lr_min}{minimum ligand receptor (features) required to keep a
cell pair(view)}
}
\value{
a list with matrices of ligand-receptors (rows) and samples
(columns) across cell type pairs (list elements)
}
\description{
Wrapper function to run `cell2cell_tensor` with LIANA output.
}
\details{
This function servers as a one-liner wrapper for factorisation
with the method described in \href{https://genomebiology.biomedcentral.com/articles/10.1186/s13059-020-02015-1}{MOFA}
}
