% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana-defaults.R
\name{liana_defaults}
\alias{liana_defaults}
\title{Function to pass Default Arguments for each method}
\usage{
liana_defaults(
  cellchat.defaults = NULL,
  connectome.defaults = NULL,
  italk.defaults = NULL,
  natmi.defaults = NULL,
  sca.defaults = NULL,
  squidpy.defaults = NULL
)
}
\arguments{
\item{cellchat.defaults}{CellChat Default Parameters \link{liana::call_cellchat}}

\item{connectome.defaults}{Connectome Default Parameters \link{liana::call_connectome}}

\item{italk.defaults}{iTALK Default Parameters \link{liana::call_italk}}

\item{natmi.defaults}{NATMI Default Parameters \link{liana::call_natmi}}

\item{sca.defaults}{SingleCellSignalR Default Parameters \link{liana::call_sca}}

\item{squidpy.defaults}{Squidpy Default Parameters \link{liana::call_squidpy}}
}
\value{
A list of the default parametres for each method
}
\description{
Function to pass Default Arguments for each method
}
\details{
Default parameters for each method can also be passed manually
}
