% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_plot.R
\name{liana_dotplot}
\alias{liana_dotplot}
\title{Liana dotplot interactions by source and target cells}
\usage{
liana_dotplot(
  liana_res,
  source_groups = NULL,
  target_groups = NULL,
  ntop = NULL,
  specificity = "natmi.edge_specificity",
  magnitude = "sca.LRscore",
  y.label = "Interactions (Ligand -> Receptor)",
  size.label = "Interaction\\nSpecificity",
  colour.label = "Expression\\nMagnitude",
  show_complex = TRUE,
  size_range = c(2, 10),
  invert_specificity = FALSE,
  invert_magnitude = FALSE,
  invert_function = function(x) -log10(x + 1e-10)
)
}
\arguments{
\item{liana_res}{aggregated `liana_wrap` results from multiple methods,
or alternatively results from running `liana_wrap` with a single method.
Should be filtered by some condition (e.g. preferential consesus ranking,
specific interactions, etc).}

\item{source_groups}{names of the source (sender) cell types (NULL = no filter)}

\item{target_groups}{names of the target cell types (NULL = no filter)}

\item{ntop}{number of interactions to return. Note that this assumes
that the tibble is sorted in descending order of interaction importance!}

\item{specificity}{column to represent the dot-size of the interaction
(by default `natmi.edge_specificity`)}

\item{magnitude}{column to represent interactions expression magnitude
(by default `sca.LRscore`)}

\item{y.label}{y label name}

\item{size.label}{size (~specificty) label name}

\item{colour.label}{colour (~magnitude) label name}

\item{show_complex}{logical whether to show complexes (default - TRUE) or
only the subunit with minimum expression.}
}
\value{
a ggplot2 object
}
\description{
Liana dotplot interactions by source and target cells
}
\details{
Here, we refer to `specificity` as how specific this interaction is
to a cell type pair regards to the rest of the cell type pairs (
e.g. CellPhoneDB's p-values, NATMI's specificity edges, Connectome's scaled weights, etc)

`magnitude` on the other hand is a direct measure of the expression alone,
by default we use SingleCellSignalR's dataset indepent LRscore (bound between 0 and 1).
Yet, one could also use CellChat's probabilities or CellPhoneDB's means, etc.
}
