% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_prep.R
\name{liana_message}
\alias{liana_message}
\title{LIANA message/warning helper function to allow for verbosity}
\usage{
liana_message(..., output = "message", verbose = TRUE)
}
\arguments{
\item{...}{zero or more objects which can be coerced to character
    (and which are pasted together with no separator) or a single
    condition object.}

\item{output}{type of output - message, warning, or stop}

\item{verbose}{logical for verbosity}
}
\description{
LIANA message/warning helper function to allow for verbosity
}
