% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_defaults.R
\name{liana_pipe.params}
\alias{liana_pipe.params}
\title{Placeholder Function to pass Default Arguments to liana_pipa and any
   derivative `get_*` method}
\usage{
liana_pipe.params()
}
\value{
A list with default arguments passed to \code{\link{liana_pipe}}
}
\description{
Placeholder Function to pass Default Arguments to liana_pipa and any
   derivative `get_*` method
}
