% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_prep.R
\name{liana_prep}
\alias{liana_prep}
\title{Function to handle different types of object as input and do basic quality checks}
\usage{
liana_prep(sce, ...)
}
\arguments{
\item{sce}{SingleCellExperiment or Seurat object}

\item{...}{dot dot dot bucket - not passed to anything, handles issues with
passing non-existing arguments}
}
\description{
Function to handle different types of object as input and do basic quality checks
}
