% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_wrapper.R
\name{liana_wrap}
\alias{liana_wrap}
\title{Run LIANA Wrapper}
\usage{
liana_wrap(
  seurat_object,
  method = c("natmi", "connectome", "logfc", "cellchat", "sca", "squidpy"),
  resource = c("OmniPath"),
  .simplify = TRUE,
  ...
)
}
\arguments{
\item{seurat_object}{seurat object}

\item{method}{method(s) to be run via liana}

\item{resource}{resource(s) to be used by the methods}

\item{.simplify}{if methods are run with only 1 resource, return a list
of tibbles for each method (default), rather than a list of lists with
method-resource combinations}

\item{...}{Pass custom method parameters to the methods via the liana wrapper
See \link{liana::liana_defaults} for more information}
}
\value{
A list of method-resource results - i.e. provided resources are run
with each method
If only one resource is selected, a single tibble (with results for that
 resource) will be returned for each of the selected methods
}
\description{
Run LIANA Wrapper
}
\details{
LIANA wrapper method that can be used to call each method with
 a given set of intercellular resources from the OmniPath universe.
Use `all` to run all resources in one go.
}
