% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{ligand_receptor_classes_bar}
\alias{ligand_receptor_classes_bar}
\title{Ligand-receptor data stacked barplots with classification}
\usage{
ligand_receptor_classes_bar(ligrec, resource, attr, largest = NULL, ...)
}
\arguments{
\item{ligrec}{List of tibbles with ligand-receptor data, as produced by
\code{\link{ligrec_overlap}}.}

\item{resource}{Character: name of the annotation resource.}

\item{attr}{Name of the classifying variable from the annotation resource
(e.g. pathway).}

\item{largest}{Numeric: how many of the largest groups to use. If `NULL`
all groups will be used.}

\item{...}{Passed to \code{\link{ligand_receptor_classes}}.}
}
\description{
Assigns classes to ligands, receptors and their connections and for each
of these entities creates a stacked barplot.
}
