% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{ligand_receptor_classes_bar_all}
\alias{ligand_receptor_classes_bar_all}
\title{Ligand-receptor data stacked barplots with classifications}
\usage{
ligand_receptor_classes_bar_all(ligrec)
}
\arguments{
\item{ligrec}{List of tibbles with ligand-receptor data, as produced by
\code{\link{ligrec_overlap}}.}
}
\value{
Returns the input data frame unchanged.
}
\description{
Calls \code{\link{ligand_receptor_classes_bar}} with classification from
many annotation resources.
}
