% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_descriptive.R
\name{ligand_receptor_upset}
\alias{ligand_receptor_upset}
\title{Upset plots of ligands, receptors and connections}
\usage{
ligand_receptor_upset(data, upset_args = list())
}
\description{
Upset plots of ligands, receptors and connections
}
