% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_descriptive.R
\name{ligrec_classes_all}
\alias{ligrec_classes_all}
\title{Ligand-receptor data stacked barplots with classifications}
\usage{
ligrec_classes_all(ligrec)
}
\arguments{
\item{ligrec}{List of tibbles with ligand-receptor data, as produced by
\code{\link{ligrec_overlap}}.}
}
\value{
Returns the input data frame unchanged.
}
\description{
Calls \code{\link{ligrec_classes}} with classification from
many annotation resources.
}
