% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_descriptive.R
\name{ligrec_decomplexify}
\alias{ligrec_decomplexify}
\title{Decomplexify specific resources that contain complexes}
\usage{
ligrec_decomplexify(ligrec)
}
\arguments{
\item{ligrec}{Nested list with ligand-receptor resources statistics, as
produced by \code{compile_ligrec}.}
}
\value{
`ligrec` contents but complexes were split into rows of subunits
   and each interaction between a complex and another protein is duplicated
   for each subunit
}
\description{
Decomplexify specific resources that contain complexes
}
