% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_descriptive.R
\name{ligrec_overheats}
\alias{ligrec_overheats}
\title{Jaccard and Shared Elements heatmaps}
\usage{
ligrec_overheats(ligrec)
}
\arguments{
\item{ligrec}{List of lists with ligand-receptor data, as produced by
\code{\link{compile_ligrec_descr}}. (to be changed to ligrec following
ligrec_overlap)}
}
\description{
Jaccard and Shared Elements heatmaps
}
