% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{ligrec_overlap}
\alias{ligrec_overlap}
\title{Overlaps between ligand-receptor resources}
\usage{
ligrec_overlap(ligrec)
}
\arguments{
\item{ligrec}{Nested list with ligand-receptor resources statistics, as
produced by \code{compile_ligrec}.}
}
\value{
The same nested list of data frames as the input (`ligrec`) with
    new columns added: unique, op_unique, omnipath, n_resources.
}
\description{
Identifies the unique and shared ligands, receptors and interactions
}
