% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_utils.R
\name{list_stats}
\alias{list_stats}
\title{Get Similarity/Dissimilarity Stats from lists with binary matrices}
\usage{
list_stats(...)
}
\arguments{
\item{...}{Any list with with DFs (typically binarized) for which we wish to
to calculate the mean, median, sd, and length.}
}
\value{
a summary tibble
}
\description{
Get Similarity/Dissimilarity Stats from lists with binary matrices
}
