% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_descriptive.R
\name{localization_examples}
\alias{localization_examples}
\title{Origin of major location attributes of intercellular interactions}
\usage{
localization_examples()
}
\description{
Origin of major location attributes of intercellular interactions
}
