% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_descriptive.R
\name{localization_ligrec_classes}
\alias{localization_ligrec_classes}
\title{Ligand-receptor classification by localization}
\usage{
localization_ligrec_classes(ligrec)
}
\arguments{
\item{ligrec}{List of tibbles with ligand-receptor data, as produced by
\code{\link{ligrec_overlap}}.}
}
\description{
Classifies ligands, receptors and their connections by localization:
either plasmam mebrane transmembrane, plasma membrane peripheral or
secreted. A receptor typically is not secreted, but here we keep this
option just to see how many of them are eventually annotated as secreted.
Such annotations are not always wrong, as some receptors indeed have
secreted forms. The connections classified by all possible combinations
of ligand and receptor localizations.
}
