% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm_funs.R
\name{map_custom}
\alias{map_custom}
\title{Helper custom map function}
\usage{
map_custom(.x, .f, parallelize, workers, ...)
}
\arguments{
\item{.x}{A list or atomic vector.}

\item{.f}{A function, formula, or vector (not necessarily atomic).

If a \strong{function}, it is used as is.

If a \strong{formula}, e.g. \code{~ .x + 2}, it is converted to a function. There
are three ways to refer to the arguments:
\itemize{
\item For a single argument function, use \code{.}
\item For a two argument function, use \code{.x} and \code{.y}
\item For more arguments, use \code{..1}, \code{..2}, \code{..3} etc
}

This syntax allows you to create very compact anonymous functions.

If \strong{character vector}, \strong{numeric vector}, or \strong{list}, it is
converted to an extractor function. Character vectors index by
name and numeric vectors index by position; use a list to index
by position and name at different levels. If a component is not
present, the value of \code{.default} will be returned.}

\item{parallelize}{whether to parallelize}

\item{workers}{Number of workers to be used in parallelization}

\item{...}{params passed to the called function and to map functions}
}
\description{
Helper custom map function
}
