% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_recomplexify.R
\name{mean0}
\alias{mean0}
\title{Helper Function which returns the mean, unless there is a 0 value then it returns 0}
\usage{
mean0(vec)
}
\arguments{
\item{vec}{numeric vector}
}
\value{
the mean of the vector unless 0 is present, then returns 0
}
\description{
Helper Function which returns the mean, unless there is a 0 value then it returns 0
}
\keyword{internal}
