% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm_funs.R
\name{mean_permute}
\alias{mean_permute}
\title{Function to calculate mean LR expression from shuffled cluster label matrices
 as done in CellPhoneDB}
\usage{
mean_permute(col_labels, sce, pb, assay.type)
}
\arguments{
\item{col_labels}{cluster labels}

\item{pb}{progress bar object}

\item{assay.type}{assay type (counts, logcounts, etc)}

\item{sce_matrix}{single cell expression matrix (transposed)}
}
\value{
Returns a list of means per gene calculated with reshuffled
   cluster/cell identity labels
}
\description{
Function to calculate mean LR expression from shuffled cluster label matrices
 as done in CellPhoneDB
}
