% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_utils.R
\docType{class}
\name{MethodSpecifics-class}
\alias{MethodSpecifics-class}
\title{S4 Class used to format benchmark output.}
\description{
S4 Class used to format benchmark output.
}
\section{Fields}{

\describe{
\item{\code{method_name}}{name of the method (e.g. CellChat)}

\item{\code{method_results}}{Named list of method-resource results}

\item{\code{method_scores}}{Named list of the measures provided by the method
and whether they should be interpreted in descending order as the value}
}}

