% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_recomplexify.R
\name{min0}
\alias{min0}
\title{Helper Function which returns the value closest to 0}
\usage{
min0(vec)
}
\arguments{
\item{vec}{numeric vector}
}
\value{
value closest to 0
}
\description{
Helper Function which returns the value closest to 0
}
