% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_recomplexify.R
\name{missing_subunits_to0}
\alias{missing_subunits_to0}
\title{Helper Function that assigns 0s to any complexes with missing subunits}
\usage{
missing_subunits_to0(lr_res, complex, entity, env)
}
\arguments{
\item{lr_res}{LR results as obtained by `lr_pipe`}

\item{complex}{complex of interest}

\item{entity}{is the complex a 'ligand' or 'receptor'}

\item{env}{environment passed to which recursive changes are saved}
}
\value{
A `lr_res` tibble with
}
\description{
Helper Function that assigns 0s to any complexes with missing subunits
}
