% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytotalk.R
\name{noise}
\alias{noise}
\title{Generate random noise}
\usage{
noise(n, noise_value = 1e-20)
}
\arguments{
\item{n}{Length of the vector to generate}
}
\value{
A vector of length n with noise in a random position
}
\description{
This function generates a vector of length n with a noise value inserted in a
random position of the vector. This is needed to compute the entropy-related
values.
}
