% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natmi_pipe.R
\name{omni_to_NATMI}
\alias{omni_to_NATMI}
\title{Reform OmniPath Resource to NATMI format and save to location}
\usage{
omni_to_NATMI(
  op_resource,
  reso_name = "placeholder",
  natmi_db_path = "input/omnipath_NATMI"
)
}
\arguments{
\item{op_resource}{Resource formatted as OmniPath}

\item{reso_name}{name of the resource}

\item{natmi_db_path}{directory in which to save the resource}
}
\description{
Reform OmniPath Resource to NATMI format and save to location
}
