% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_omnipath.R
\name{omnipath_intercell}
\alias{omnipath_intercell}
\title{Retrieves intercellular interactions from OmniPath}
\usage{
omnipath_intercell(quality = NULL, ligrec = FALSE)
}
\arguments{
\item{quality}{Quality measures for OmniPath CompositeDB (biased due to the
prevalence of e.g. Ramilowski in the resources)}

\item{ligrec}{whether to filter according to side}
}
\value{
A tibble with Intercell interactions from OmnIPath
}
\description{
Retrieves intercellular interactions from OmniPath
}
