% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_omnipath.R
\name{omnipath_partners}
\alias{omnipath_partners}
\title{Retrieves intercellular communication partners (transmitters or receivers)
from OmniPath}
\usage{
omnipath_partners(side, quality = NULL, ligrec = FALSE)
}
\arguments{
\item{side}{Either 'ligand' (trans), 'receptor' (rec) or 'both' (both
short or long notation can be used).}

\item{quality}{Numeric: a quality threshold for OmniPath between 0 and 1.}

\item{ligrec}{Use only ligand-receptor interactions from OmniPath.}
}
\description{
Retrieves intercellular communication partners (transmitters or receivers)
from OmniPath
}
