% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{order_by_group_size}
\alias{order_by_group_size}
\title{Orders the levels of a factor variable by their number of elements}
\usage{
order_by_group_size(data, var)
}
\arguments{
\item{data}{A data frame.}

\item{var}{Name of a variable in the data frame.}
}
\value{
The data frame with the variable converted to a factor and its
    levels ordered.
}
\description{
Orders the levels of a factor variable by their number of elements
}
