% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_utils.R
\name{plot_abundance_summary}
\alias{plot_abundance_summary}
\title{Function to Plot Abundance Summary}
\usage{
plot_abundance_summary(ctqc, ncol = 3)
}
\arguments{
\item{ctqc}{cell type quality control summary obtained from
`get_abundance_summary`}

\item{ncol}{number of columns for the facet wrap}
}
\value{
a ggplot2 object
}
\description{
Function to Plot Abundance Summary
}
\keyword{internal}
