% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_tensor.R
\name{plot_context_boxplot}
\alias{plot_context_boxplot}
\title{Generate boxplots with significance}
\usage{
plot_context_boxplot(sce, group_col, test = "t.test", ...)
}
\arguments{
\item{sce}{SingleCellExperiment with factors output from tensor-cell2cell}

\item{group_col}{context descriptor - to be obtained from `colData(sce)`}

\item{...}{arguments passed to the test used.}
}
\description{
Generate boxplots with significance
}
