% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_tensor.R
\name{plot_contexts_umap}
\alias{plot_contexts_umap}
\title{Function to plot a UMAP of context loadings}
\usage{
plot_contexts_umap(factors, key_sep = "[|]", ...)
}
\arguments{
\item{factors}{factors output from tensor-cell2cel formatted with
`format_c2c_factors`}

\item{key_sep}{separator for the sample and condition names. These are
used to jointly name the contexts - technically the names of the
`context_df_dict` list passed to the `liana_tensor_c2c` function.}
}
\value{
a ggplot2 object
}
\description{
Function to plot a UMAP of context loadings
}
