% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plot_freq_pca}
\alias{plot_freq_pca}
\title{PCA plot for Cell type/cluster pair frequency for 'significant/top' hits}
\usage{
plot_freq_pca(sig_list)
}
\arguments{
\item{sig_list}{named list of significant hits. Named list of methods with
each element being a named list of resources}
}
\value{
A ggplot2 object
}
\description{
PCA plot for Cell type/cluster pair frequency for 'significant/top' hits
}
