% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plotSaveUset}
\alias{plotSaveUset}
\title{Helper function to save Upset plot}
\usage{
plotSaveUset(upset_df, dir)
}
\arguments{
\item{upset_df}{prepForUpset output}

\item{dir}{path to save upset plot}
}
\value{
Null
}
\description{
Helper function to save Upset plot
}
