% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_roc.R
\name{prepare_for_roc}
\alias{prepare_for_roc}
\title{Helper function used to format LR results to the appropriate format for the
calc_curve function}
\usage{
prepare_for_roc(lr_form_res, ground, predictor_metric)
}
\arguments{
\item{lr_form_res}{Formatted LR results with subsampling and the resources
used for each tool.}

\item{ground}{"Ground Truth" - i.e. results considered significant
when using the whole dataset (no subsampling)}
}
\value{
A tidy data frame used to assess robustness
}
\description{
Helper function used to format LR results to the appropriate format for the
calc_curve function
}
