% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{prepForUpset}
\alias{prepForUpset}
\title{Helper function to prepare list for UpsetPlot}
\usage{
prepForUpset(named_list)
}
\arguments{
\item{named_list}{a named list with sugnificant LR results}
}
\value{
Matrix/DF of 0 and 1 where 1 means that the interaction
is present and 0 means that it is not
}
\description{
Helper function to prepare list for UpsetPlot
}
