% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_aggregate.R
\name{rank_method}
\alias{rank_method}
\title{Helper function to rank each method}
\usage{
rank_method(liana_res, method_name, mode = "specificity")
}
\arguments{
\item{liana_res}{liana_results for a single method}

\item{method_name}{name of the method}

\item{mode}{ranking to be carried out. Accepted modes are `specificity`
and `magnitude`. The first is meant to reflect the specificity of interactions
across all cell types, while the latter typically reflects how highly expressed
is a given interaction.}
}
\description{
Helper function to rank each method
}
\details{
this function makes use of liana's `liana:::.score_specs` and
`liana:::.score_housekeep` functions.
}
