% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_aggregate.R
\docType{class}
\name{RankSpecifics-class}
\alias{RankSpecifics-class}
\title{S4 Class used to generate aggregate/consesus scores for the methods.}
\description{
S4 Class used to generate aggregate/consesus scores for the methods.
}
\section{Fields}{

\describe{
\item{\code{method_name}}{name of the method (e.g. cellchat)}

\item{\code{method_score}}{The interaction score provided by the method (typically
the score that reflects the specificity of interaction)}

\item{\code{descending_order}}{whether the score should be interpreted in
descending order (i.e. highest score for an interaction is most likely)}
}}

