% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_defaults.R
\name{reassign_params}
\alias{reassign_params}
\title{Helper function to replace default parameters}
\usage{
reassign_params(replacements, defaults)
}
\arguments{
\item{replacements}{named list corresponding to default arguments}

\item{defaults}{named list with default arguments}
}
\value{
a named list with the same arguments
}
\description{
Helper function to replace default parameters
}
