% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_ortho.R
\name{recode.character2}
\alias{recode.character2}
\title{Modified `dplyr::recode` function}
\usage{
\method{recode}{character2}(.x, ..., .default = NULL, .missing = NULL, .missing_fun)
}
\arguments{
\item{.x}{A vector to modify}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Replacements. For character and factor \code{.x}, these should be named
and replacement is based only on their name. For numeric \code{.x}, these can be
named or not. If not named, the replacement is done based on position i.e.
\code{.x} represents positions to look for in replacements. See examples.

When named, the argument names should be the current values to be replaced, and the
argument values should be the new (replacement) values.

All replacements must be the same type, and must have either
length one or the same length as \code{.x}.}

\item{.default}{If supplied, all values not otherwise matched will
be given this value. If not supplied and if the replacements are
the same type as the original values in \code{.x}, unmatched
values are not changed. If not supplied and if the replacements
are not compatible, unmatched values are replaced with \code{NA}.

\code{.default} must be either length 1 or the same length as
\code{.x}.}

\item{.missing}{If supplied, any missing values in \code{.x} will be
replaced by this value. Must be either length 1 or the same length as
\code{.x}.}

\item{.missing_fun}{Function to modify any missing homologs/strings
`NULL` by default and any missing values will be discarded.
For example, one could be set it to `str_to_title` to format all symbols, or
any other format in a scenario where a homolog dictionary is not available
for the organism of interest.}
}
\description{
Modified `dplyr::recode` function
}
\details{
enables to modify unmatched genesymbols
}
\keyword{internal}
