% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_omnipath.R
\name{reform_omni}
\alias{reform_omni}
\title{Helper Function to Reformat Omni_resources for LR Pipeline}
\usage{
reform_omni(omni_resources)
}
\arguments{
\item{omni_resources}{List: output from \code{compile_ligrec}.}
}
\description{
Helper Function to Reformat Omni_resources for LR Pipeline
}
