% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_utils.R
\name{reform_rank_frequencies}
\alias{reform_rank_frequencies}
\title{Helper function to convert list with all resources ranked to frequencies df}
\usage{
reform_rank_frequencies(frequencies_list)
}
\arguments{
\item{frequencies_list}{list with all resources ranked to frequencies df}
}
\description{
Helper function to convert list with all resources ranked to frequencies df
}
