% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_descriptive.R
\name{replace_inf}
\alias{replace_inf}
\title{Replace infinite values with a value slightly lower or higher than the
extrema}
\usage{
replace_inf(data, var, s = 1.1)
}
\description{
Replace infinite values with a value slightly lower or higher than the
extrema
}
