% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bench_robust.R
\name{robust_format_res}
\alias{robust_format_res}
\title{Helper function used to format robustness results}
\usage{
robust_format_res(lr_res, db_list, subsampling, .res_order = TRUE)
}
\arguments{
\item{lr_res}{Unformatted robustness results}

\item{db_list}{list of the OmniPath resources/dbs used in the subsampling}

\item{subsampling}{Vector of values used in the subsampling}

\item{.res_order}{bool whether the results are ordered by resource and
subsampling (e.g. OmniPath 1, OmniPath 0.8, OmniPath 0.6)}
}
\value{
A formatted tidy tibble with DB names and subsampling
}
\description{
Helper function used to format robustness results
}
