% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bench_robust.R
\name{robust_get_roc}
\alias{robust_get_roc}
\title{Function used to calculate the ROC curve of each robustness run}
\usage{
robust_get_roc(lr_form_res, predictor_metric, predictor_thresh, .rank = FALSE)
}
\arguments{
\item{lr_form_res}{formatted LR results returned by robust_format_res()}

\item{predictor_metric}{Name of the predictor metric used in the
corresponding LR algorithm}

\item{predictor_thresh}{Threshold for the predictor metric
(e.g. 0.05 for pvalue)}

\item{.rank}{bool: whether to rank statistic by percentage}
}
\value{
A Formatted tibble with calculated Robustness ROC
}
\description{
Function used to calculate the ROC curve of each robustness run
}
