% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_roc.R
\name{robust_roc_plot}
\alias{robust_roc_plot}
\title{Get Robustness Roc Plot}
\usage{
robust_roc_plot(lr_format_roc)
}
\arguments{
\item{lr_format_roc}{formatted LR results with ROC calculated}
}
\value{
ggplot2 object ROC plot
}
\description{
Get Robustness Roc Plot
}
