% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sca_pipe.R
\name{sca_formatDB}
\alias{sca_formatDB}
\title{Helper Function to convert Omni to LRdb Format}
\usage{
sca_formatDB(op_resource)
}
\arguments{
\item{op_resource}{OmniPath resource}
}
\description{
Helper Function to convert Omni to LRdb Format
}
