% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bench_robust.R
\name{seurat_subsample}
\alias{seurat_subsample}
\title{Function to subsample Seurat object to fraction of cells by type}
\usage{
seurat_subsample(
  seurat_object,
  subsampling = 0.8,
  .idents = "seurat_clusters",
  .seed = 1004
)
}
\arguments{
\item{seurat_object}{Processed Seurat object with sc data}

\item{subsampling}{Value used to subsample cell clusters}

\item{.idents}{name of the idents}

\item{.seed}{Integer used to set the RNG seed}
}
\value{
A suerat object subsampled with cells subsampled by cluster
}
\description{
Function to subsample Seurat object to fraction of cells by type
}
