% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_pipe.R
\name{seurat_to_sce}
\alias{seurat_to_sce}
\title{Helper function to convert from Seurat to SCE.}
\usage{
seurat_to_sce(seurat_object, entity_genes, assay)
}
\arguments{
\item{seurat_object}{seurat_obect}

\item{entity_genes}{ligand and receptor genes (i.e. genes that are in the
CCC resource)}

\item{assay}{assay}
}
\description{
Helper function to convert from Seurat to SCE.
}
\details{
LIANA uses the SingleCellExperiment/Bioconductor architecture to
generate the LR summary
}
