% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_descriptive.R
\name{shorten_resources}
\alias{shorten_resources}
\title{In a data frame shortens the names of certain resources}
\usage{
shorten_resources(data)
}
\description{
In a data frame shortens the names of certain resources
}
