% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_ortho.R
\name{show_homologene}
\alias{show_homologene}
\title{Helper function to show available organisms via OmnipathR's homologene resource}
\usage{
show_homologene()
}
\description{
Helper function to show available organisms via OmnipathR's homologene resource
}
