% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_wrapper.R
\name{show_methods}
\alias{show_methods}
\title{Helper Function to return the methods in LIANA}
\usage{
show_methods()
}
\description{
Helper Function to return the methods in LIANA
}
\details{
methods starting with `call_*` were re-implemented in liana and
   albeit their original pipelines (and packages are still supported),
   we recommend using the liana re-implementations for efficiency
}
