% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_utils.R
\name{simdist_resmet}
\alias{simdist_resmet}
\title{Get (Dis)Similarities per Resource/Method Combinations}
\usage{
simdist_resmet(sig_list, ...)
}
\arguments{
\item{sig_list}{list of top hits}

\item{...}{
  Arguments passed on to \code{\link[proxy:dist]{proxy::simil}}
  \describe{
    \item{\code{x}}{For \code{dist} and \code{simil}, a numeric matrix object, a data frame, or a list. A vector
    will be converted into a column matrix. For \code{as.simil} and
    \code{as.dist}, an object of class \code{dist} and
    \code{simil}, respectively, or a numeric matrix. For
    \code{pr_dist2simil} and \code{pr_simil2dist}, any numeric vector.}
    \item{\code{y}}{\code{NULL}, or a similar object than \code{x}}
    \item{\code{method}}{a function, a registry entry, or a mnemonic string referencing the
    proximity measure. A list of all available measures can be obtained
    using \code{\link[proxy]{pr_DB}} (see examples). The default for \code{dist} is
    \code{"Euclidean"}, and for \code{simil} \code{"correlation"}.}
    \item{\code{diag}}{logical value indicating whether the diagonal of the
    distance/similarity matrix should be printed by
    \code{\link[proxy]{print.dist}}/\code{\link[proxy]{print.simil}}. Note that the
    diagonal values are never stored in \code{dist} objects. 

    In the context of \code{as.matrix} the value to use on the diagonal
    representing self-proximities. In case of similarities, this
    defaults to \code{NA} since a priori there are no upper bounds, so
    the maximum similarity needs to be specified by the user.}
    \item{\code{upper}}{logical value indicating whether the upper triangle of the
    distance/similarity matrix should be printed by
    \code{\link[proxy]{print.dist}}/\code{\link[proxy]{print.simil}}}
    \item{\code{pairwise}}{logical value indicating whether distances should be
    computed for the pairs of \code{x} and \code{y} only.}
    \item{\code{by_rows}}{logical indicating whether proximities between rows, or
    columns should be computed.}
    \item{\code{convert_distances}}{logical indicating
    whether distances should be automatically converted into
    similarities (and the other way round) if needed.}
    \item{\code{auto_convert_data_frames}}{logical indicating whether data frames
    should be converted to matrices if all variables are numeric,
    or all are logical, or all are complex.}
  }}
}
\description{
Get (Dis)Similarities per Resource/Method Combinations
}
