% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squidpy_pipe.R
\name{squidpy_reformat}
\alias{squidpy_reformat}
\title{Helper function to reformat squidpy function r}
\usage{
squidpy_reformat(.name, .pval_list, .mean_list)
}
\arguments{
\item{.name}{omnipath resource name}

\item{.pval_list}{p-value results from different dbs as a list from squidpy}

\item{.mean_list}{mean list from squidpy}
}
\description{
Helper function to reformat squidpy function r
}
