% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NATMI_pipe.R
\name{str_split_helper}
\alias{str_split_helper}
\title{Split and format strings for subsampling}
\usage{
str_split_helper(path, project_name)
}
\arguments{
\item{path}{path to CSV (em/annotations) to split and format to the
current subsampling taken from a seurat object project name}

\item{project_name}{seurat object project name}
}
\value{
Path to save subsampling EM and Annotations
}
\description{
Split and format strings for subsampling
}
