% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_descriptive.R
\name{summarize_overlaps}
\alias{summarize_overlaps}
\title{Summary of overlaps between ligand-receptor resources}
\usage{
summarize_overlaps(ligrec_olap)
}
\arguments{
\item{ligrec_olap}{Nested list with ligand-receptor resources statistics
with the shared and unique entities identified. Produced by
\code{ligrec_overlap}.}
}
\value{
Summarized counts of unique and shared elements for each resource.
}
\description{
Counts the unique and shared ligands, receptors and interactions
}
