% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_utils.R
\name{top_enh}
\alias{top_enh}
\title{Helper Function to handle specific cases for CellChat and Squidpy}
\usage{
top_enh(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[dplyr:top_n]{dplyr::top_n}}
  \describe{
    \item{\code{x}}{A data frame.}
    \item{\code{n}}{Number of rows to return for \code{top_n()}, fraction of rows to
return for \code{top_frac()}. If \code{n} is positive, selects the top rows.
If negative, selects the bottom rows.
If \code{x} is grouped, this is the number (or fraction) of rows per group.
Will include more rows if there are ties.}
    \item{\code{wt}}{(Optional). The variable to use for ordering. If not
specified, defaults to the last variable in the tbl.}
  }}
}
\value{
Ordered tibble/df as from top_n
}
\description{
Helper Function to handle specific cases for CellChat and Squidpy
}
