% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_descriptive.R
\name{total_unique_bar}
\alias{total_unique_bar}
\title{Total vs. unique barplot}
\usage{
total_unique_bar(ligrec_olap)
}
\arguments{
\item{ligrec_olap}{Summarized ligand-receptor overlaps, as produced by
\code{summarize_overlaps}.}
}
\description{
Creates a barplot with the unique, shared and total number of ligands,
receptors and connections in each resource.
}
