% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_descriptive.R
\name{uniq_per_res}
\alias{uniq_per_res}
\title{Function to Calculate Unique Interactions Per Resource by Category}
\usage{
uniq_per_res(ligrec_olap)
}
\arguments{
\item{ligrec_olap}{Nested list with ligand-receptor resources statistics
with the shared and unique entities identified. Produced by
\code{ligrec_overlap}.}
}
\description{
Function to Calculate Unique Interactions Per Resource by Category
}
