% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_descriptive.R
\name{upset_generic}
\alias{upset_generic}
\title{Creates an upset plot}
\usage{
upset_generic(data, label, omnipath, upset_args, ...)
}
\arguments{
\item{data}{A data frame, one element from the output of
\code{ligrec_overlap}.}

\item{label}{A label for the file name. Should refer to the entities we
classify e.g. "ligand".}

\item{omnipath}{Logical: whether to include OmniPath.}

\item{upset_args}{List: additional arguments for \code{UpSetR::upset}.}

\item{...}{Column names: the first one should be the set assignment, the
rest define unique entities.}
}
\description{
A wrapper to create an upset plot from a data frame
}
