% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_descriptive.R
\name{binarize_resources}
\alias{binarize_resources}
\title{Helper Function to Convert a List of Resources into a binarized DF}
\usage{
binarize_resources(interaction_list)
}
\arguments{
\item{interaction_list}{list resources with interactions alone}
}
\value{
Returns a 1/0 dataframe where 1 is assigned to interactions
   present in a given resource, 0 to absent interactions
}
\description{
Helper Function to Convert a List of Resources into a binarized DF
}
