% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellchat_pipe.R
\name{call_cellchat}
\alias{call_cellchat}
\title{CellChat x Omni Pipe function}
\usage{
call_cellchat(
  op_resource,
  seurat_object,
  .format = TRUE,
  exclude_anns = c(),
  nboot = 100,
  assay = "RNA",
  .seed = 1004,
  .normalize = FALSE,
  .do_parallel = FALSE,
  .raw_use = TRUE,
  ...
)
}
\arguments{
\item{op_resource}{OmniPath Intercell Resource DN}

\item{seurat_object}{Seurat object as input}

\item{.format}{bool whether to format output}

\item{exclude_anns}{Annotation criteria to be excluded}

\item{nboot}{number of bootstraps to calculate p-value}

\item{.normalize}{# bool whether to normalize non-normalized data with}

\item{.raw_use}{whether use the raw data or gene expression data pojectected
to a ppi}

\item{...}{
  Arguments passed on to \code{\link[CellChat:subsetCommunication]{CellChat::subsetCommunication}}
  \describe{
    \item{\code{object}}{CellChat object}
    \item{\code{net}}{Alternative input is a data frame with at least with three columns defining the cell-cell communication network ("source","target","interaction_name")}
    \item{\code{slot.name}}{the slot name of object: slot.name = "net" when extracting the inferred communications at the level of ligands/receptors; slot.name = "netP" when extracting the inferred communications at the level of signaling pathways}
    \item{\code{sources.use}}{a vector giving the index or the name of source cell groups}
    \item{\code{targets.use}}{a vector giving the index or the name of target cell groups.}
    \item{\code{signaling}}{a character vector giving the name of signaling pathways of interest}
    \item{\code{pairLR.use}}{a data frame consisting of one column named either "interaction_name" or "pathway_name", defining the interactions of interest}
    \item{\code{thresh}}{threshold of the p-value for determining significant interaction}
    \item{\code{datasets}}{select the inferred cell-cell communications from a particular `datasets` when inputing a data frame `net`}
    \item{\code{ligand.pvalues}}{set threshold for ligand genes

ligand.pvalues: threshold for pvalues in the differential expression gene analysis (DEG)

ligand.logFC: threshold for logFoldChange in the DEG analysis; When ligand.logFC > 0, keep upgulated genes; otherwise, kepp downregulated genes

ligand.pct.1: threshold for the percent of expressed genes in the defined 'positive' cell group. keep genes with percent greater than ligand.pct.1

ligand.pct.2: threshold for the percent of expressed genes in the cells except for the defined 'positive' cell group}
    \item{\code{ligand.logFC}}{set threshold for ligand genes

ligand.pvalues: threshold for pvalues in the differential expression gene analysis (DEG)

ligand.logFC: threshold for logFoldChange in the DEG analysis; When ligand.logFC > 0, keep upgulated genes; otherwise, kepp downregulated genes

ligand.pct.1: threshold for the percent of expressed genes in the defined 'positive' cell group. keep genes with percent greater than ligand.pct.1

ligand.pct.2: threshold for the percent of expressed genes in the cells except for the defined 'positive' cell group}
    \item{\code{ligand.pct.1}}{set threshold for ligand genes

ligand.pvalues: threshold for pvalues in the differential expression gene analysis (DEG)

ligand.logFC: threshold for logFoldChange in the DEG analysis; When ligand.logFC > 0, keep upgulated genes; otherwise, kepp downregulated genes

ligand.pct.1: threshold for the percent of expressed genes in the defined 'positive' cell group. keep genes with percent greater than ligand.pct.1

ligand.pct.2: threshold for the percent of expressed genes in the cells except for the defined 'positive' cell group}
    \item{\code{ligand.pct.2}}{set threshold for ligand genes

ligand.pvalues: threshold for pvalues in the differential expression gene analysis (DEG)

ligand.logFC: threshold for logFoldChange in the DEG analysis; When ligand.logFC > 0, keep upgulated genes; otherwise, kepp downregulated genes

ligand.pct.1: threshold for the percent of expressed genes in the defined 'positive' cell group. keep genes with percent greater than ligand.pct.1

ligand.pct.2: threshold for the percent of expressed genes in the cells except for the defined 'positive' cell group}
    \item{\code{receptor.pvalues}}{set threshold for receptor genes}
    \item{\code{receptor.logFC}}{set threshold for receptor genes}
    \item{\code{receptor.pct.1}}{set threshold for receptor genes}
    \item{\code{receptor.pct.2}}{set threshold for receptor genes}
  }}
}
\value{
A DF of intercellular communication network
}
\description{
CellChat x Omni Pipe function
}
\details{
CellChat's objects are not documented/exported thus the
  whole package has to be imported.
}
