% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellchat_pipe.R
\name{call_cellchat}
\alias{call_cellchat}
\title{CellChat x Omni Pipe function}
\usage{
call_cellchat(
  op_resource,
  seurat_object,
  .format = TRUE,
  exclude_anns = c(),
  nboot = 100,
  assay = "RNA",
  .seed = 1004,
  .normalize = FALSE,
  .do_parallel = FALSE,
  .raw_use = TRUE,
  ...
)
}
\arguments{
\item{op_resource}{OmniPath Intercell Resource DN}

\item{seurat_object}{Seurat object as input}

\item{.format}{bool whether to format output}

\item{exclude_anns}{Annotation criteria to be excluded}

\item{nboot}{number of bootstraps to calculate p-value}

\item{.normalize}{# bool whether to normalize non-normalized data with}

\item{.raw_use}{whether use the raw data or gene expression data pojectected
to a ppi}
}
\value{
A DF of intercellular communication network
}
\description{
CellChat x Omni Pipe function
}
\details{
CellChat's objects are not documented/exported thus the
   whole package has to be imported
}
