% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectome_pipe.R
\name{call_connectome}
\alias{call_connectome}
\title{Function to call connectome with databases from OmniPath}
\usage{
call_connectome(seurat_object, op_resource = NULL, .format = TRUE, ...)
}
\arguments{
\item{seurat_object}{Seurat object as input}

\item{op_resource}{OmniPath Intercell Resource DN}

\item{.format}{bool whether to format output}

\item{...}{dot params passed to connectome}
}
\value{
An unfiltered connectome results df
}
\description{
Function to call connectome with databases from OmniPath
}
\details{
Stats:
1) The ‘weight_norm’ edge attribute is derived from the normalized expression
 of the ligand and the receptor in the single-cell data.
2) The ‘weight_scale’ edge attribute is derived from the z-scores of the ligand
 and the receptor in each edge, and is of higher value when the ligand and receptor
  are more specific to a given pair of cell types
3) DEG p-values for L and R
}
