% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iTALK_pipe.R
\name{call_italk}
\alias{call_italk}
\title{Run iTALK with OmniPath data}
\usage{
call_italk(
  op_resource,
  seurat_object,
  assay = "SCT",
  .format = TRUE,
  .DE = FALSE,
  .deg = NULL,
  ...
)
}
\arguments{
\item{op_resource}{OmniPath Intercell Resource DN}

\item{seurat_object}{Seurat object as input}

\item{assay}{assay to use from Seurat object}

\item{.format}{bool: whether to format output}

\item{.DE}{bool: whether to use DE (TRUE) or highlyVarGenes (FALSE)}

\item{.deg}{if is NULL run FindAllMarkers}
}
\value{
An unfiltered iTALK df sorted by relevance
}
\description{
Run iTALK with OmniPath data
}
\details{
Stats:
Ligand and Receptor Expressions (and P-values if ran with .DE==TRUE)
Dot params are inherited from Seurat::FindAllMarkers, if .deg = TRUE
}
