% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iTALK_pipe.R
\name{call_italk}
\alias{call_italk}
\title{Function to call iTalk with databases from OmniPath}
\usage{
call_italk(
  op_resource,
  seurat_object,
  assay = "SCT",
  .format = TRUE,
  .DE = FALSE,
  .deg = NULL,
  ...
)
}
\arguments{
\item{op_resource}{OmniPath Intercell Resource DN}

\item{seurat_object}{Seurat object as input}

\item{assay}{assay to use from Seurat object}

\item{.format}{bool: whether to format output}

\item{.DE}{bool: whether to use DE (TRUE) or highlyVarGenes (FALSE)}

\item{.deg}{if is NULL run FindAllMarkers}

\item{...}{
  Arguments passed on to \code{\link[iTALK:rawParse]{iTALK::rawParse}}
  \describe{
    \item{\code{data}}{Input data, raw or normalized count with 'cell_type' column}
    \item{\code{top_genes}}{(scale 1 to 100) Top percent highly expressed genes used
to find ligand-receptor pairs, default is 50}
    \item{\code{stats}}{Whether calculates the mean or the median of the data. Available
options are 'mean' and 'median'.}
  }}
}
\value{
An unfiltered iTALK df sorted by relevance
}
\description{
Function to call iTalk with databases from OmniPath
}
\details{
Stats:
Ligand and Receptor Expressions (and P-values if ran with .DE==TRUE)
Dot params are inherited from Seurat::FindAllMarkers, if .deg = TRUE
}
