% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/italk_pipe.R
\name{call_italk}
\alias{call_italk}
\title{Run iTALK with OmniPath data [[DEPRECATED]]}
\usage{
call_italk(sce, op_resource, assay = "RNA", .format = TRUE, ...)
}
\arguments{
\item{sce}{Seurat object or SingleCellExperiment as input}

\item{op_resource}{OmniPath Intercell Resource DN}

\item{assay}{assay to use from Seurat object}

\item{.format}{bool: whether to format output}

\item{...}{Parameters passed to Seurat FindMarkers (ref requires import)}
}
\value{
An unfiltered iTALK df sorted by relevance

In this case, we use the product of the logFC rather than thresholding, as in
the original implementation.
}
\description{
Run iTALK with OmniPath data [[DEPRECATED]]
}
\details{
In order to be comparable with the remainder of the methods, we
calculate the mean of the ligand and receptor logFC.
The original implementation only uses the DE genes above a certain logFC
threshold.
}
