% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NATMI_pipe.R
\name{call_natmi}
\alias{call_natmi}
\title{Call NATMI Pipeline from R with Resources Querried from OmniPath}
\usage{
call_natmi(
  op_resource,
  seurat_object,
  expr_file = "em.csv",
  meta_file = "metadata.csv",
  output_dir = "NATMI_test",
  assay = "RNA",
  num_cor = 4,
  .format = TRUE,
  .write_data = TRUE,
  .seed = 1004,
  .natmi_path = NULL
)
}
\arguments{
\item{op_resource}{List of OmniPath resources}

\item{seurat_object}{Seurat object}

\item{expr_file}{expression matrix file name}

\item{meta_file}{annotations (i.e. clusters) file name}

\item{output_dir}{NATMI output directory}

\item{assay}{Seurat assay to be used}

\item{.format}{bool whether to format output}

\item{.write_data}{bool whether Extract data from Seurat Object}

\item{.natmi_path}{path of NATMI code and dbs (by default set to liana path)}

\item{omnidbs_dir}{path of saved omnipath resources}

\item{.default_run}{bool whether to run default DBs or not}
}
\value{
DF with NATMI results
}
\description{
Call NATMI Pipeline from R with Resources Querried from OmniPath
}
\details{
This function will save NATMI dbs folder, then it will call the
    NATMI Python from the NATMI dir and save the output into a specified
    directory in NATMI's path.
    It will then load the csvs and format the output to a list of lists.

    By default, NATMI's path is set to that of LIANA, but any alternative
    path can be passed

==============================================================================
NATMI Arguments:
  --interDB INTERDB
                        lrc2p (default) has literature supported ligand-receptor pairs | lrc2a has putative and literature supported ligand-receptor pairs | the user-supplied interaction database can also be used by calling the name of database file without extension
  --interSpecies INTERSPECIES
                        human (default) | mouse | expandp | expanda
  --emFile EMFILE       the path to the normalised expression matrix file with row names (gene identifiers) and column names (cell-type/single-cell identifiers)
  --annFile ANNFILE     the path to the metafile in which column one has single-cell identifiers and column two has corresponding cluster IDs (see file 'toy.sc.ann.txt' as an example)
  --species SPECIES     human (default) | mouse | rat | zebrafish | fruitfly | chimpanzee | dog | monkey | cattle | chicken | frog | mosquito | nematode | thalecress | rice | riceblastfungus | bakeryeast | neurosporacrassa | fissionyeast | eremotheciumgossypii | kluyveromyceslactis, 21 species are supported
  --idType IDTYPE       symbol (default) | entrez(https://www.ncbi.nlm.nih.gov/gene) | ensembl(https://www.ensembl.org/) | uniprot(https://www.uniprot.org/) | hgnc(https://www.genenames.org/) | mgi(http://www.informatics.jax.org/mgihome/nomen/index.shtml) | custom(gene identifier used in the expression matrix)
  --coreNum CORENUM     the number of CPU cores used, default is one
  --out OUT             the path to save the analysis results
(Taken From NATMI's GitHub Page)

Stats:
1) The mean-expression edge weights
2) The specificity-based edge weights
* a weight of 1 means both the ligand and receptor are only expressed
 in one cell type
}
