% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sca_pipe.R
\name{call_sca}
\alias{call_sca}
\title{Function to call SingleCellSignalR with databases from OmniPath [[DEPRECATED]]}
\usage{
call_sca(
  sce,
  op_resource,
  .format = TRUE,
  assay = "RNA",
  assay.type = "logcounts",
  ...
)
}
\arguments{
\item{sce}{SingleCellExperiment or SeuratObject as input}

\item{op_resource}{OmniPath Intercell Resource DN}

\item{.format}{bool whether to format output}

\item{assay}{Seurat assay data to use}

\item{assay.type}{count slot (logcounts by default)}
}
\value{
An unfiltered SCA tibble
}
\description{
Function to call SingleCellSignalR with databases from OmniPath [[DEPRECATED]]
}
\details{
Stats:
LRScore = sqrt(LR product)/mean(raw counts) * sqrt(LR product) where
expression of l > 0 and r > 0
LRScore = 1 is the highest (~ most likely hit), 0 is the lowest.
}
