% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squidpy_pipe.R
\name{call_squidpy}
\alias{call_squidpy}
\title{Call Squidpy Pipeline via reticulate with OmniPath and format results}
\usage{
call_squidpy(seurat_object, op_resource, .seed = 1004, conda_env = NULL, ...)
}
\arguments{
\item{seurat_object}{Seurat object as input}

\item{op_resource}{Tibble or list of OmniPath resources, typically obtained via
\code{\link{select_resource}}}

\item{.seed}{used to python seed}

\item{conda_env}{python conda environment to run Squidpy; set to liana_env by default}

\item{...}{kwargs passed to Squidpy; For more information see:
\link{https://squidpy.readthedocs.io/en/latest/api/squidpy.gr.ligrec.html#squidpy.gr.ligrec}}
}
\value{
A list of Squidpy results for each resource
}
\description{
Call Squidpy Pipeline via reticulate with OmniPath and format results
}
\details{
CellPhoneDBv2 algorithm re-implementation in Python

Note that `cluster_key` is a parameter passed to the Squidpy function,
by default this will be set to the default Ident of the Seurat object.
}
