% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squidpy_pipe.R
\name{call_squidpyR}
\alias{call_squidpyR}
\title{Call Squidpy Pipeline via reticulate with OmniPath and format results}
\usage{
call_squidpyR(seurat_object, op_resource, .seed = 1004, ...)
}
\arguments{
\item{seurat_object}{Seurat object as input}

\item{op_resource}{List of OmniPath resources}

\item{.seed}{used to python seed}

\item{...}{kwargs passed to Squidpy; For more information see:
\https://squidpy.readthedocs.io/en/latest/api/squidpy.gr.ligrec.html#squidpy.gr.ligrec}
}
\value{
A list of Squidpy results for each resource
}
\description{
Call Squidpy Pipeline via reticulate with OmniPath and format results
}
\details{
CellPhoneDB v2 algorithm implementation in Python
}
