% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_descriptive.R
\name{classes_enrich}
\alias{classes_enrich}
\title{Enrichment dot plot from a data frame of classified entities}
\usage{
classes_enrich(data, entity, resource, var, ...)
}
\arguments{
\item{data}{A data frame with classified entities (ligands, receptors or
interactions).}

\item{entity}{The name of the entity, to be included in the output file
name and the y axis label.}

\item{resource}{The name of the resource, to be included in the output
file name.}

\item{var}{Name of the classifying variable.}

\item{...}{Passed to \code{\link{enrich2}}.}
}
\value{
Returns `NULL`.
}
\description{
Enrichment dot plot from a data frame of classified entities
}
