% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_omnipath.R
\name{compile_ligrec}
\alias{compile_ligrec}
\title{Function to get unfiltered intercell resources
For each resource and OmniPath variant compiles tables of ligands,
receptors and interactions}
\usage{
compile_ligrec()
}
\description{
Function to get unfiltered intercell resources
For each resource and OmniPath variant compiles tables of ligands,
receptors and interactions
}
\details{
calls on omnipath_intercell, intercell_connections, get_partners,
and intercell_connections
}
