% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_omnipath.R
\name{compile_ligrec}
\alias{compile_ligrec}
\title{Function to get unfiltered intercell resources
For each resource and OmniPath variant compiles tables of ligands,
receptors and interactions}
\usage{
compile_ligrec(lr_pipeline = TRUE)
}
\arguments{
\item{lr_pipeline}{bool whether to format for lr_pipeline and remove

duplicate LRs (mainly from composite OmniDB due to category (adhesion vs lr))}
}
\value{
A list of OmniPath resources formatted according to the method pipes
}
\description{
Function to get unfiltered intercell resources
For each resource and OmniPath variant compiles tables of ligands,
receptors and interactions
}
\details{
calls on omnipath_intercell, intercell_connections, get_partners,
and intercell_connections
}
\keyword{internal}
