% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_call.R
\name{cpdb_permute}
\alias{cpdb_permute}
\title{Function to calculate mean LR expression from shuffled cluster label matrices
 as done in CellPhoneDB}
\usage{
cpdb_permute(col_labels, sce_mat, trim)
}
\arguments{
\item{col_labels}{cluster labels}

\item{trim}{truncate ends of mean}

\item{sce_matrix}{single cell expression matrix (transposed)}

\item{lr_og}{LR res with only relevant}
}
\value{
Returns a list of means per gene calculated with reshuffled
   cluster/cell identity labels
}
\description{
Function to calculate mean LR expression from shuffled cluster label matrices
 as done in CellPhoneDB
}
