% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_call.R
\name{cpdb_score}
\alias{cpdb_score}
\title{Function to calculate p-values as in CellPhoneDB}
\usage{
cpdb_score(lr_res, perm_means, parallelize, workers, score_col = "pvalue")
}
\arguments{
\item{lr_res}{liana_pipe results}

\item{parallelize}{whether to parallelize}

\item{workers}{Number of workers to be used in parallelization}

\item{score_col}{name of the score column}
}
\description{
Function to calculate p-values as in CellPhoneDB
}
