% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstalk_funs.R
\name{cytotalk_score}
\alias{cytotalk_score}
\title{Compute cross-talk score from a Seurat Object (DEPRECATED)}
\usage{
cytotalk_score(lr_res, sce, score_col, assay.type, seed, ...)
}
\arguments{
\item{lr_res}{liana_pipe results}

\item{sce}{SingleCellExperiment object}

\item{score_col}{name of the score column}

\item{assay.type}{data slot to be used}

\item{seed}{random seed}

\item{...}{placeholder}
}
\description{
Compute cross-talk score from a Seurat Object (DEPRECATED)
}
\details{
This function executes all the required functions to extract
and compute the cross-talk scores as defined by the Cytotalk authors.
}
\keyword{internal}
