% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_descriptive.R
\name{decomplexify}
\alias{decomplexify}
\title{Helper Function to 'decomplexify' a column or vector of columns in a resource}
\usage{
decomplexify(resource, column)
}
\arguments{
\item{resource}{a ligrec resource}

\item{column}{column to separate and pivot long (e.g. genesymbol)}
}
\value{
returns a longer tibble with complex subunits on seperate rows
}
\description{
Helper Function to 'decomplexify' a column or vector of columns in a resource
}
