% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_pipe.R
\name{decomplexify}
\alias{decomplexify}
\title{Helper Function to 'decomplexify' ligands and receptors into}
\usage{
decomplexify(resource, columns = c("source_genesymbol", "target_genesymbol"))
}
\arguments{
\item{resource}{a ligrec resource}

\item{columns}{columns to separate and pivot long (e.g. genesymbol or uniprot)}
}
\value{
returns a longer tibble with complex subunits on seperate rows
}
\description{
Helper Function to 'decomplexify' ligands and receptors into
}
