% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_aggregate.R
\name{.aggregate_rank}
\alias{.aggregate_rank}
\title{Robust Aggregate ranks using `RobustRankAggreg`}
\usage{
.aggregate_rank(liana_mlist, ...)
}
\arguments{
\item{liana_mlist}{liana list with method tibbles}

\item{...}{Parametres passed to `aggregateRanks` from `RobustRankAggreg`}
}
\description{
Robust Aggregate ranks using `RobustRankAggreg`
}
