% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_wrapper.R
\name{.filt_liana_pipe}
\alias{.filt_liana_pipe}
\title{Helper Function to do prop filtering of liana_pipe output}
\usage{
.filt_liana_pipe(liana_res, method, ...)
}
\arguments{
\item{liana_res}{resutls from `liana_pipe`}

\item{method}{current method}

\item{...}{
  Arguments passed on to \code{\link[=liana_defaults]{liana_defaults}}
  \describe{
    \item{\code{liana_pipe.params}}{list of Parameters passed to NATMI \code{\link{liana_pipe}}}
    \item{\code{liana_call.params}}{list of Parameters passed to NATMI \code{\link{liana_call}}}
    \item{\code{cellchat.params}}{list of Parameters passed to CellChat \code{\link{call_cellchat}}}
    \item{\code{squidpy.params}}{list of Parameters passed to Squidpy \code{\link{call_squidpy}}}
    \item{\code{call_connectome.params}}{list of Parameters passed to Connectome \code{\link{call_connectome}}}
    \item{\code{call_italk.params}}{list of Parameters passed to iTALK \code{\link{call_italk}}}
    \item{\code{call_natmi.params}}{list of Parameters passed to NATMI \code{\link{call_natmi}}}
    \item{\code{call_sca.params}}{list of Parameters passed to SingleCellSignalR \code{\link{call_sca}}}
    \item{\code{assay}}{Assay name passed to `call_italk`, `call_sca`, `call_cellchat`,
and `call_connectome`}
    \item{\code{expr_prop}}{minimum proportion of gene expression per cell type (0.1 by default).
This is further applied at a method-specific level - it is not applied to
 Connectome and Cytotalk, as these methods calculate/normalize the score (or
 statistics used in their scores) across all cell types or cell type pairs.}
    \item{\code{seed}}{random seed integer}
    \item{\code{parallelize}}{whether to parallelize cellphonedb-like}
    \item{\code{workers}}{number of workers to be called}
  }}
}
\description{
Helper Function to do prop filtering of liana_pipe output
}
