% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_prep.R
\name{.format_idents}
\alias{.format_idents}
\title{Helper Function to get/format the required indentity if required}
\usage{
.format_idents(metadata, active_idents, idents_col, object_class, verbose)
}
\description{
Helper Function to get/format the required indentity if required
}
