% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_pipe.R
\name{.get_invcounts}
\alias{.get_invcounts}
\title{Helper function to generate inversed counts (i.e. normalized but not logged)}
\usage{
.get_invcounts(sce, assay.type)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{assay.type}{counts slot}
}
\description{
Helper function to generate inversed counts (i.e. normalized but not logged)
}
