% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NATMI_pipe.R
\name{FormatNatmi}
\alias{FormatNatmi}
\title{Helper function to load NATMI results from folder and format appropriately}
\usage{
FormatNatmi(output_path, .format = TRUE)
}
\arguments{
\item{output_path}{NATMI output path}

\item{.format}{bool whether to format output}
}
\value{
A list of NATMI results per resource loaded from the output directory
}
\description{
Helper function to load NATMI results from folder and format appropriately
}
