% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_tensor.R
\name{generate_lr_geneset}
\alias{generate_lr_geneset}
\title{Generate a geneset resource for each LR}
\usage{
generate_lr_geneset(sce, resource, lr_sep = "^")
}
\arguments{
\item{sce}{SingleCellExperiment object with liana_tensor_c2c computed}

\item{resource}{resource with `source`, `target`, `weight` columns}
}
\value{
a tibble in decoupleR format
}
\description{
Generate a geneset resource for each LR
}
