% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_utils.R
\name{generate_orthologs}
\alias{generate_orthologs}
\title{Function to translate Human genes to organism X orthologues}
\usage{
generate_orthologs(op_resource, symbols_dict)
}
\arguments{
\item{op_resource}{resource in the format of OmniPath/LIANA}

\item{symbols_dict}{dictionary (named list) with human genesymbols and ortholog
in a second species}
}
\description{
Function to translate Human genes to organism X orthologues
}
