% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_utils.R
\name{get_bc_stats}
\alias{get_bc_stats}
\title{Function to get Similarity/Dissimilarity stats from a lits of top hits}
\usage{
get_bc_stats(sig_list)
}
\arguments{
\item{sig_list}{list of top hits per method-resource combo}
}
\value{
A named list of  similarity/dissimilarity stats
}
\description{
Function to get Similarity/Dissimilarity stats from a lits of top hits
}
\details{
returns the results for the same method with different resources
  as well as results for the same resource with different methods.
  Furthermore, it returns summary stats for each of these which include:
  mean and st dev of  /similarity/dissimilarities per method (and resource)
  as well as the mean of means and the mean of standard deviations
}
