% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{get_binary_frequencies}
\alias{get_binary_frequencies}
\title{Get binary activity frequencies per cell-cell pair}
\usage{
get_binary_frequencies(sig_list)
}
\arguments{
\item{sig_list}{named list of significant hits. Named list of methods with
each element being a named list of resources}
}
\value{
A tibble of cell pair frequencies, based on binarized activity
}
\description{
Get binary activity frequencies per cell-cell pair
}
