% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_call.R
\name{get_connectome}
\alias{get_connectome}
\title{Function to obtain connectome-like weights}
\usage{
get_connectome(lr_res, ...)
}
\arguments{
\item{lr_res}{ligand-receptor DE results and other stats between clusters}

\item{...}{
  Arguments passed on to \code{\link[=liana_call]{liana_call}}
  \describe{
    \item{\code{op_resource}}{resource tibble obtained via \link{liana::select_resource}}
  }}
}
\value{
Returns a tibble with specificity weights (`weight_sc`) as calculated
   by Connectome
}
\description{
Function to obtain connectome-like weights
}
