% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_call.R
\name{get_connectome}
\alias{get_connectome}
\title{Function to obtain connectome-like weights}
\usage{
get_connectome(lr_res, ...)
}
\arguments{
\item{lr_res}{ligand-receptor DE results and other stats between clusters}

\item{...}{
  Arguments passed on to \code{\link[=liana_call]{liana_call}}
  \describe{
    \item{\code{seurat_object}}{Seurat object}
    \item{\code{op_resource}}{resource tibble obtained via \link{liana::select_resource}}
    \item{\code{decomplexify}}{whether to dissociate complexes into subunits and hence
and hence take complexes into account (decomplexify) or not}
    \item{\code{complex_policy}}{policy how to account for the presence of complexes.
Following the example of \href{https://squidpy.readthedocs.io/en/stable/api/squidpy.gr.ligrec.html}{Squidpy} valid options are:
`'min0'` (default): select the subunit with the change/expression closest to 0
 (as in \href{https://github.com/Teichlab/cellphonedb}{CellPhoneDB} when working with expression)
`'all'`: returns all possible subunits separately;
Alternatively, pass any other base or user-defined function that reduces a vector to a single number (e.g. mean, min, etc)}
  }}
}
\value{
Returns a tibble with specificity weights (`weight_sc`) as calculated
   by Connectome
}
\description{
Function to obtain connectome-like weights
}
