% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_call.R
\name{get_logfc}
\alias{get_logfc}
\title{Function to obtain logFC weights}
\usage{
get_logfc(lr_res, ...)
}
\arguments{
\item{lr_res}{ligand-receptor DE results and other stats between clusters}

\item{...}{
  Arguments passed on to \code{\link[=liana_call]{liana_call}}
  \describe{
    \item{\code{op_resource}}{resource tibble obtained via \link{liana::select_resource}}
  }}
}
\value{
Returns a tibble with a logFC metric (`logfc_comb`). `logfc_comb` is
   calculated as the product of the (1 vs the rest) log2FC for each ligand
   and receptor gene
}
\description{
Function to obtain logFC weights
}
