% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_call.R
\name{get_logfc}
\alias{get_logfc}
\title{Function to obtain connectome-like weights}
\usage{
get_logfc(seurat_object, op_resource, ...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=.liana_call]{.liana_call}}
  \describe{
    \item{\code{seurat_object}}{Seurat object}
    \item{\code{op_resource}}{resource tibble obtained via \link{liana::select_resource}}
    \item{\code{decomplexify}}{whether to decomplexify or not}
    \item{\code{lr_res}}{ligand-receptor DE results and other stats between clusters}
  }}
}
\value{
lr_res modified to be method-specific
}
\description{
Function to obtain connectome-like weights
}
