% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_call.R
\name{get_natmi}
\alias{get_natmi}
\title{Function to obtain NATMI-like weights}
\usage{
get_natmi(seurat_object, op_resource, ...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=.liana_call]{.liana_call}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
Returns a tibble with specificity weights (`edge_specificity`)
   as calculated by natmi
}
\description{
Function to obtain NATMI-like weights
}
