% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_omnipath.R
\name{get_omni_resources}
\alias{get_omni_resources}
\title{Function to call OmniPath resources}
\usage{
get_omni_resources()
}
\value{
A list with intercellular resources from OmniPath
}
\description{
Function to call OmniPath resources
}
