% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_omnipath.R
\name{get_partners}
\alias{get_partners}
\title{Retrieves intercellular communication partners (ligands or receptors) from
one ligand-receptor resource.}
\usage{
get_partners(side, resource, ...)
}
\arguments{
\item{side}{'ligand' (trans), 'receptor' (rec) or 'both' (both short or long notation can be used)}

\item{resource}{Name of current resource (taken from get_lr_resources)}

\item{...}{
  Arguments passed on to \code{\link[=omnipath_intercell]{omnipath_intercell}}
  \describe{
    \item{\code{quality}}{Quality measures for OmniPath CompositeDB (biased due to the
prevalence of e.g. Ramilowski in the resources)}
    \item{\code{ligrec}}{whether to filter according to side}
  }}
}
\description{
Retrieves intercellular communication partners (ligands or receptors) from
one ligand-receptor resource.
}
