% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_omnipath.R
\name{get_partners}
\alias{get_partners}
\title{Retrieves intercellular communication partners (ligands or receptors) from
one ligand-receptor resource.}
\usage{
get_partners(side, resource, ...)
}
\arguments{
\item{side}{'ligand' (trans), 'receptor' (rec) or 'both' (both short or
long notation can be used)}

\item{resource}{Name of current resource (taken from get_lr_resources)}

\item{...}{Inherit dot params from \link{OmnipathR::omnipath_intercell}}
}
\description{
Retrieves intercellular communication partners (ligands or receptors) from
one ligand-receptor resource.
}
