% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_utils.R
\name{get_rank_frequencies}
\alias{get_rank_frequencies}
\title{Convert list with MethodSpecifics objects to a dataframe with ranked
cell_pair frequencies.}
\usage{
get_rank_frequencies(spec_list)
}
\arguments{
\item{spec_list}{list with appropriately populated MethodSpecifics objects}
}
\value{
a tibble with cell_pair frequencies represented by the ranked
normalized scores for each method by cell_pair
}
\description{
Convert list with MethodSpecifics objects to a dataframe with ranked
cell_pair frequencies.
}
\details{
See format_rank_frequencies for details
}
