% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_omnipath.R
\name{get_receptors}
\alias{get_receptors}
\title{Retrieves receptors from one ligand-receptor resource}
\usage{
get_receptors(resource, ...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=intercell_connections]{intercell_connections}}
  \describe{
    \item{\code{resource}}{Name of current resource (taken from get_lr_resources)}
  }}
}
\description{
Retrieves receptors from one ligand-receptor resource
}
