% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_descriptive.R
\name{hgnc_ligrec_classes}
\alias{hgnc_ligrec_classes}
\title{Ligand-receptor classification from HGNC}
\usage{
hgnc_ligrec_classes(ligrec, largest = 15)
}
\arguments{
\item{ligrec}{List of tibbles with ligand-receptor data, as produced by
\code{\link{ligrec_overlap}}.}

\item{largest}{Numeric: how many of the largest groups to use. If `NULL`
all groups will be used.}
}
\description{
Ligand-receptor classification from HGNC
}
