% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_omnipath.R
\name{intercell_connections}
\alias{intercell_connections}
\title{Retrieves the interactions from one ligand-receptor resource}
\usage{
intercell_connections(resource, ...)
}
\description{
Retrieves the interactions from one ligand-receptor resource
}
