% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_omnipath.R
\name{intercell_connections}
\alias{intercell_connections}
\title{Retrieves the interactions from one ligand-receptor resource}
\usage{
intercell_connections(resource, ...)
}
\arguments{
\item{resource}{Name of current resource (taken from get_lr_resources)}

\item{...}{
  Arguments passed on to \code{\link[OmnipathR:import_post_translational_interactions]{OmnipathR::import_post_translational_interactions}}
  \describe{
    \item{\code{resources}}{interactions not reported in these databases are
removed. See \code{\link[OmnipathR]{get_interaction_resources}} for more information.}
    \item{\code{organism}}{Interactions are available for human, mouse and rat.
Choose among: 9606 human (default), 10116 rat and 10090 Mouse}
    \item{\code{exclude}}{datasets to exclude}
    \item{\code{references_by_resource}}{if FALSE, removes the resource name prefixes
from the references (PubMed IDs); this way the information which reference
comes from which resource will be lost and the PubMed IDs will be unique.}
  }}
}
\description{
Retrieves the interactions from one ligand-receptor resource
}
