% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_descriptive.R
\name{jacc_pairwise}
\alias{jacc_pairwise}
\title{Pairwise Jaccard Index between resources}
\usage{
jacc_pairwise(ligrec_binary)
}
\value{
Returns a long Dataframe with Jaccard Indeces
}
\description{
Pairwise Jaccard Index between resources
}
