% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_call.R
\name{liana_call}
\alias{liana_call}
\title{Wrapper Function to obtain scores via liana_pipe}
\usage{
liana_call(method, lr_res, ...)
}
\arguments{
\item{method}{name of the method to be called}

\item{lr_res}{ligand-receptor DE results and other stats between clusters}

\item{...}{
  Arguments passed on to \code{\link[=liana_pipe]{liana_pipe}}
  \describe{
    \item{\code{sce}}{SingleCellExperiment Object}
    \item{\code{op_resource}}{resource tibble obtained via \link{liana::select_resource}}
    \item{\code{assay}}{assay to be used ("RNA" by default)}
    \item{\code{assay.type}}{- the type of data to be used to calculate the means
(logcounts by default), available options are: "counts" and "logcounts"}
    \item{\code{verbose}}{logical for verbosity}
    \item{\code{cell.adj}}{cell adjacency tibble/dataframe /w weights by which we will
`multiply` the relevant columns. Any cell pairs with a weights of 0 will be
filtered out.
Note that if working with LIANA's default methods, we suggest weights >= 0 & =< 1.
This ensure that all methods' score will be meaningfully weighed without
changing the interpretation of their scores, thus allow one to filter SCA,
rank NATMI, etc.}
    \item{\code{test.type}}{String specifying the type of pairwise test to perform -
a t-test with \code{"t"}, a Wilcoxon rank sum test with \code{"wilcox"}, 
or a binomial test with \code{"binom"}.}
    \item{\code{pval.type}}{A string specifying how p-values are to be combined across pairwise comparisons for a given group/cluster.}
    \item{\code{base}}{base for conversion from log-tranformed ~CPM back to ~CPM.}
  }}
}
\value{
lr_res modified to be method-specific
}
\description{
Wrapper Function to obtain scores via liana_pipe
}
