% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_defaults.R
\name{liana_defaults}
\alias{liana_defaults}
\title{Function to pass Default Arguments for each method}
\usage{
liana_defaults(
  assay = "RNA",
  assay.type = "logcounts",
  expr_prop = 0.1,
  seed = 1004,
  complex_policy = "mean0",
  parallelize = FALSE,
  workers = 8,
  permutation.params = NULL,
  liana_pipe.params = NULL,
  liana_call.params = NULL,
  cellphonedb.params = NULL,
  natmi.params = NULL,
  sca.params = NULL,
  connectome.params = NULL,
  cytotalk.params = NULL,
  logfc.params = NULL,
  scconnect.params = NULL,
  cellchat.params = NULL,
  squidpy.params = NULL,
  call_sca.params = NULL,
  call_natmi.params = NULL,
  call_connectome.params = NULL,
  call_italk.params = NULL
)
}
\arguments{
\item{assay}{Assay name passed to `call_italk`, `call_sca`, `call_cellchat`,
and `call_connectome`}

\item{expr_prop}{minimum proportion of gene expression per cell type (0.1 by default).
This is further applied at a method-specific level - it is not applied to
 Connectome and Cytotalk, as these methods calculate/normalize the score (or
 statistics used in their scores) across all cell types or cell type pairs.}

\item{seed}{random seed integer}

\item{parallelize}{whether to parallelize cellphonedb-like}

\item{workers}{number of workers to be called}

\item{liana_pipe.params}{list of Parameters passed to NATMI \code{\link{liana_pipe}}}

\item{liana_call.params}{list of Parameters passed to NATMI \code{\link{liana_call}}}

\item{cellchat.params}{list of Parameters passed to CellChat \code{\link{call_cellchat}}}

\item{squidpy.params}{list of Parameters passed to Squidpy \code{\link{call_squidpy}}}

\item{call_sca.params}{list of Parameters passed to SingleCellSignalR \code{\link{call_sca}}}

\item{call_natmi.params}{list of Parameters passed to NATMI \code{\link{call_natmi}}}

\item{call_connectome.params}{list of Parameters passed to Connectome \code{\link{call_connectome}}}

\item{call_italk.params}{list of Parameters passed to iTALK \code{\link{call_italk}}}

\item{trim}{the fraction (0 to 0.5) of observations to be trimmed from each
end of x before the mean is computed. This is relevant only for the
CellPhoneDB algorithm re-implementation in liana.}
}
\value{
A list of the default parameters for each method
}
\description{
Function to pass Default Arguments for each method
}
\details{
The default parameters for each method can also be overwritten by
 manually passing a list of parameters for the appropraite method
  \code{\link{liana_wrap}}

Further, each `get_*` method will by default obtain the default params passed
   via \code{\link{liana_pipe}} and \code{\link{liana_call}}. This is done so that most steps
   required for the calculation of these methods are undertaken only once.
}
\examples{
# get a `named` list with all default parameters passed to liana.
def_params <- liana_defaults()
# any of these can then be overwritten and passed to `...` in `liana_wrap`
# with the `.params` suffix to the parameter name type. For example,
liana_res <- liana_wrap(testdata,
                        permutation.params = list(parms=2),
                        liana_pipe.params = list(test.type='wilcox'))
}
