% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana-defaults.R
\name{liana_defaults}
\alias{liana_defaults}
\title{Function to pass Default Arguments for each method}
\usage{
liana_defaults(
  assay = "RNA",
  cellchat.params = NULL,
  connectome.params = NULL,
  italk.params = NULL,
  natmi.params = NULL,
  sca.params = NULL,
  squidpy.params = NULL
)
}
\arguments{
\item{cellchat.params}{CellChat Parameters \link{liana::call_cellchat}}

\item{connectome.params}{Connectome Parameters \link{liana::call_connectome}}

\item{italk.params}{iTALK Parameters \link{liana::call_italk}}

\item{natmi.params}{NATMI Parameters \link{liana::call_natmi}}

\item{sca.params}{SingleCellSignalR Parameters \link{liana::call_sca}}

\item{squidpy.params}{Squidpy Parameters \link{liana::call_squidpy}}
}
\value{
A list of the default parameters for each method
}
\description{
Function to pass Default Arguments for each method
}
\details{
The default parameters for each method can also be overwritten by
 manually passing a list of parameters for the appropraite method
  \link{liana::liana_wrap}
}
