% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_defaults.R
\name{liana_defaults}
\alias{liana_defaults}
\title{Function to pass Default Arguments for each method}
\usage{
liana_defaults(
  assay = "RNA",
  decomplexify = FALSE,
  cellchat.params = NULL,
  squidpy.params = NULL,
  sca.params = NULL,
  liana_pipe.params = NULL,
  liana_call.params = NULL,
  call_natmi.params = NULL,
  call_connectome.params = NULL,
  call_italk.params = NULL
)
}
\arguments{
\item{assay}{Assay name passed to `call_italk`, `call_sca`, `call_cellchat`,
and `call_connectome`}

\item{decomplexify}{specify whether complexes in the resource should be
dissociated and taken into account}

\item{cellchat.params}{list of Parameters passed to CellChat \code{\link{call_cellchat}}}

\item{squidpy.params}{list of Parameters passed to Squidpy \code{\link{call_squidpy}}}

\item{sca.params}{list of Parameters passed to SingleCellSignalR \code{\link{call_sca}}}

\item{liana_pipe.params}{list of Parameters passed to NATMI \code{\link{liana_pipe}}}

\item{liana_call.params}{list of Parameters passed to NATMI \code{\link{liana_call}}}

\item{call_natmi.params}{list of Parameters passed to NATMI \code{\link{call_natmi}}}

\item{call_connectome.params}{list of Parameters passed to Connectome \code{\link{call_connectome}}}

\item{call_italk.params}{list of Parameters passed to iTALK \code{\link{call_italk}}}
}
\value{
A list of the default parameters for each method
}
\description{
Function to pass Default Arguments for each method
}
\details{

}
