% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_defaults.R
\name{liana_defaults}
\alias{liana_defaults}
\title{Function to pass Default Arguments for each method}
\usage{
liana_defaults(
  assay = "RNA",
  assay.type = "logcounts",
  decomplexify = TRUE,
  expr_prop = 0.1,
  seed = 1004,
  trim = 0,
  parallelize = FALSE,
  workers = 8,
  cellchat.params = NULL,
  squidpy.params = NULL,
  call_sca.params = NULL,
  cellphonedb.params = NULL,
  cytotalk.params = NULL,
  permutation.params = NULL,
  liana_pipe.params = NULL,
  liana_call.params = NULL,
  call_natmi.params = NULL,
  call_connectome.params = NULL,
  call_italk.params = NULL
)
}
\arguments{
\item{assay}{Assay name passed to `call_italk`, `call_sca`, `call_cellchat`,
and `call_connectome`}

\item{decomplexify}{specify whether complexes in the resource should be
dissociated and taken into account}

\item{expr_prop}{minimum proportion of gene expression per cell type (0.2 by default).
One should consider setting this to an appropriate value between 0 and 1,
as an assumptions of these methods is that communication is coordinated at the cluster level.}

\item{seed}{random seed integer}

\item{trim}{the fraction (0 to 0.5) of observations to be trimmed from each
end of x before the mean is computed. This is relevant only for the
CellPhoneDB algorithm re-implementation in liana.}

\item{parallelize}{whether to parallelize cellphonedb-like}

\item{workers}{number of workers to be called}

\item{cellchat.params}{list of Parameters passed to CellChat \code{\link{call_cellchat}}}

\item{squidpy.params}{list of Parameters passed to Squidpy \code{\link{call_squidpy}}}

\item{call_sca.params}{list of Parameters passed to SingleCellSignalR \code{\link{call_sca}}}

\item{liana_pipe.params}{list of Parameters passed to NATMI \code{\link{liana_pipe}}}

\item{liana_call.params}{list of Parameters passed to NATMI \code{\link{liana_call}}}

\item{call_natmi.params}{list of Parameters passed to NATMI \code{\link{call_natmi}}}

\item{call_connectome.params}{list of Parameters passed to Connectome \code{\link{call_connectome}}}

\item{call_italk.params}{list of Parameters passed to iTALK \code{\link{call_italk}}}
}
\value{
A list of the default parameters for each method
}
\description{
Function to pass Default Arguments for each method
}
\details{
The default parameters for each method can also be overwritten by
 manually passing a list of parameters for the appropraite method
  \code{\link{liana_wrap}}

Further, each `get_*` method will by default obtain the default params passed
   via \code{\link{liana_pipe}} and \code{\link{liana_call}}. This is done so that most steps
   required for the calculation of these methods are undertaken only once.
}
