% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_plot.R
\name{liana_dotplot}
\alias{liana_dotplot}
\title{Liana dotplot interactions by source and target cells}
\usage{
liana_dotplot(
  liana_agg,
  source_groups,
  target_groups,
  specificity = "natmi.edge_specificity",
  magnitude = "sca.LRscore",
  y.label = "Interactions (Ligand -> Receptor)",
  size.label = "Interaction\\nSpecificity",
  colour.label = "Expression\\nMagnitude",
  show_complex = TRUE
)
}
\arguments{
\item{liana_agg}{aggregated `liana_wrap` results -> preferentially filtered
by some condition (e.g. preferential ranking, specific interactions, etc)}

\item{source_groups}{names of the source(sender) cell types}

\item{target_groups}{names of the target cell types}

\item{specificity}{column to represent the dot-size of the interaction
(by default `natmi.edge_specificity`)}

\item{magnitude}{column to represent interactions expression magnitude
(by default `sca.LRscore`)}

\item{y.label}{y label name}

\item{size.label}{size (~specificty) label name}

\item{colour.label}{colour (~magnitude) label name}

\item{show_complex}{logical whether to show complexes (default - TRUE) or
only the subunit with minimum expression.}
}
\value{
a ggplot2 object
}
\description{
Liana dotplot interactions by source and target cells
}
\details{
Here, we refer to `specificity` as how specific this interaction is
to a cell type pair regards to the rest of the cell type pairs (
e.g. CellPhoneDB's p-values, NATMI's specificity edges, Connectome's scaled weights, etc)

`magnitude` on the other hand is a direct measure of the expression alone,
by default we use SingleCellSignalR's dataset indepent LRscore (bound between 0 and 1).
Yet, one could also use CellChat's probabilities or CellPhoneDB's means, etc.
}
