% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_plot.R
\name{liana_dotplot}
\alias{liana_dotplot}
\title{Liana dotplot interactions by source and target cells}
\usage{
liana_dotplot(
  liana_agg,
  source_groups,
  target_groups,
  specificity = "natmi.edge_specificity",
  show_complex = FALSE,
  magnitude = "sca.LRscore"
)
}
\arguments{
\item{liana_agg}{aggregated `liana_wrap` results -> preferentially filtered
by some condition (e.g. preferential ranking, specific interactions, etc)}

\item{source_groups}{names of the source(sender) cell types}

\item{target_groups}{names of the target cell types}

\item{specificity}{column to represent the specificity of the interaction}

\item{magnitude}{column to represent the magnitude of interaction (by default
'sca.LRscore')}
}
\value{
a ggplot2 object
}
\description{
Liana dotplot interactions by source and target cells
}
\details{
Here, we refer to `specificity` as how specific this interaction is
to a cell type pair regards to the rest of the cell type pairs (
e.g. CellPhoneDB's p-values, NATMI's specificity edges, Connectome's scaled weights, etc)

`magnitude` on the other hand is a direct measure of the expression alone,
by default we use SingleCellSignalR's dataset indepent LRscore (bound between 0 and 1).
Yet, one could also use CellChat's probabilities or CellPhoneDB's means, etc.
}
