% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_pipe.R
\name{liana_pipe}
\alias{liana_pipe}
\title{Liana Pipe which runs DE analysis and merges needed information for LR inference}
\usage{
liana_pipe(
  sce,
  op_resource,
  test.type = "wilcox",
  pval.type = "all",
  assay = "RNA",
  assay.type = "logcounts",
  verbose = TRUE,
  base,
  cell.adj = NULL
)
}
\arguments{
\item{sce}{SingleCellExperiment Object}

\item{op_resource}{resource tibble obtained via \link{liana::select_resource}}

\item{test.type}{String specifying the type of pairwise test to perform -
a t-test with \code{"t"}, a Wilcoxon rank sum test with \code{"wilcox"}, 
or a binomial test with \code{"binom"}.}

\item{pval.type}{A string specifying how p-values are to be combined across pairwise comparisons for a given group/cluster.}

\item{assay}{assay to be used ("RNA" by default)}

\item{assay.type}{- the type of data to be used to calculate the means
(logcounts by default), available options are: "counts" and "logcounts"}

\item{verbose}{logical for verbosity}

\item{base}{base for conversion}

\item{cell.adj}{cell adjacency tibble/dataframe /w weights by which we will
`multiply` the relevant columns. Note that if working with LIANA's default
methods, we suggest weights >= 0 & =< 1. This ensure that all methods' score
will be meaningfully weighed without changing the interpretation of their
scores, thus allow one to filter SCA, rank NATMI, etc.}
}
\value{
Returns a tibble with information required for LR calculations downstream
}
\description{
Liana Pipe which runs DE analysis and merges needed information for LR inference
}
