% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_pipe.R
\name{liana_pipe}
\alias{liana_pipe}
\title{Liana Pipe which runs DE analysis and merges needed information for LR inference}
\usage{
liana_pipe(
  sce,
  op_resource,
  decomplexify = TRUE,
  test.type = "wilcox",
  pval.type = "all",
  trim = 0,
  assay = "RNA",
  assay.type = "logcounts"
)
}
\arguments{
\item{sce}{SingleCellExperiment Object}

\item{op_resource}{resource tibble obtained via \link{liana::select_resource}}

\item{decomplexify}{whether to dissociate complexes into subunits and hence
and hence take complexes into account (decomplexify) or not}

\item{test.type}{String specifying the type of pairwise test to perform -
a t-test with \code{"t"}, a Wilcoxon rank sum test with \code{"wilcox"}, 
or a binomial test with \code{"binom"}.}

\item{pval.type}{A string specifying how p-values are to be combined across pairwise comparisons for a given group/cluster.}

\item{assay}{assay to be used ("RNA" by default)}

\item{assay.type}{- the type of data to be used to calculate the means
(logcounts by default), available options are: "counts" and "logcounts"}

\item{expr_prop}{minimum proportion of gene expression per cell type (0.2 by default),
yet one should consider setting this to an appropriate value between 0 and 1,
as an assumptions of these method is that communication is coordinated at the cluster level.}
}
\value{
Returns a tibble with information required for LR calculations downstream
}
\description{
Liana Pipe which runs DE analysis and merges needed information for LR inference
}
