% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_pipe.R
\name{liana_pipe}
\alias{liana_pipe}
\title{Liana Pipe which runs DE analysis and merges needed information for LR inference}
\usage{
liana_pipe(
  seurat_object,
  op_resource,
  decomplexify = TRUE,
  test.type = "t",
  pval.type = "all",
  seed = 1234
)
}
\arguments{
\item{seurat_object}{Seurat object}

\item{op_resource}{resource tibble obtained via \link{liana::select_resource}}

\item{test.type}{`test.type` passed to \link{scran::findMarkers}}

\item{pval.type}{`pval.type` passed to \link{scran::findMarkers}}

\item{seed}{Set Random Seed}
}
\value{
Returns a tibble with information required for LR calc
}
\description{
Liana Pipe which runs DE analysis and merges needed information for LR inference
}
