% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_pipe.R
\name{liana_pipe}
\alias{liana_pipe}
\title{Liana Pipe which runs DE analysis and merges needed information for LR inference}
\usage{
liana_pipe(
  seurat_object,
  op_resource,
  decomplexify = TRUE,
  test.type = "t",
  pval.type = "all",
  expr_prop = 0,
  trim = 0.1
)
}
\arguments{
\item{seurat_object}{Seurat object}

\item{op_resource}{resource tibble obtained via \link{liana::select_resource}}

\item{decomplexify}{whether to dissociate complexes into subunits and hence
and hence take complexes into account (decomplexify) or not}

\item{test.type}{String specifying the type of pairwise test to perform -
a t-test with \code{"t"}, a Wilcoxon rank sum test with \code{"wilcox"}, 
or a binomial test with \code{"binom"}.}

\item{pval.type}{A string specifying how p-values are to be combined across pairwise comparisons for a given group/cluster.}

\item{expr_prop}{minimum proportion of gene expression per cell type (0 by default),
yet perhaps one should consider setting this to an appropriate value between 0 and 1,
as an assumptions of these method is that communication is coordinated at the cluster level.}

\item{seed}{Set Random Seed}
}
\value{
Returns a tibble with information required for LR calc
}
\description{
Liana Pipe which runs DE analysis and merges needed information for LR inference
}
