% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_prep.R
\name{liana_prep}
\alias{liana_prep}
\title{Function to handle different types of object as input and do basic quality checks}
\usage{
liana_prep(sce, ...)
}
\arguments{
\item{sce}{Seurat or SingleCellExperiment object}
}
\description{
Function to handle different types of object as input and do basic quality checks
}
