% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_call.R
\name{liana_scores}
\alias{liana_scores}
\title{Function to obtain different scoring schemes}
\usage{
liana_scores(score_object, lr_res, decomplexify = TRUE, ...)
}
\arguments{
\item{score_object}{score_object specific to the test obtained from score_specs}

\item{lr_res}{ligand-receptor DE results and other stats between clusters}

\item{decomplexify}{whether to dissociate complexes into subunits and hence
and hence take complexes into account (decomplexify) or not}
}
\value{
lr_res modified to be method-specific
}
\description{
Function to obtain different scoring schemes
}
