% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_wrapper.R
\name{liana_wrap}
\alias{liana_wrap}
\title{LIANA wrapper function}
\usage{
liana_wrap(
  seurat_object,
  method = c("natmi", "connectome", "logfc", "sca", "cellphonedb"),
  resource = c("OmniPath"),
  external_resource,
  .simplify = TRUE,
  ...
)
}
\arguments{
\item{seurat_object}{seurat object}

\item{method}{method(s) to be run via liana}

\item{resource}{resource(s) to be used by the methods (Use `all` to run all resources in one go),
or `custom` to run liana_wrap with an appropriately formatted custom resource, passed via `exernal_resource`}

\item{external_resource}{external resource in OmniPath tibble format}

\item{.simplify}{if methods are run with only 1 resource, return a list
of tibbles for each method (default), rather than a list of lists with
method-resource combinations}

\item{...}{
  Arguments passed on to \code{\link[=liana_defaults]{liana_defaults}}
  \describe{
    \item{\code{liana_pipe.params}}{list of Parameters passed to NATMI \code{\link{liana_pipe}}}
    \item{\code{liana_call.params}}{list of Parameters passed to NATMI \code{\link{liana_call}}}
    \item{\code{cellchat.params}}{list of Parameters passed to CellChat \code{\link{call_cellchat}}}
    \item{\code{squidpy.params}}{list of Parameters passed to Squidpy \code{\link{call_squidpy}}}
    \item{\code{call_connectome.params}}{list of Parameters passed to Connectome \code{\link{call_connectome}}}
    \item{\code{call_italk.params}}{list of Parameters passed to iTALK \code{\link{call_italk}}}
    \item{\code{call_natmi.params}}{list of Parameters passed to NATMI \code{\link{call_natmi}}}
    \item{\code{call_sca.params}}{list of Parameters passed to SingleCellSignalR \code{\link{call_sca}}}
    \item{\code{assay}}{Assay name passed to `call_italk`, `call_sca`, `call_cellchat`,
and `call_connectome`}
    \item{\code{expr_prop}}{minimum proportion of gene expression per cell type (0.1 by default).
This is further applied at a method-specific level - it is not applied to
 Connectome and Cytotalk, as these methods calculate/normalize the score (or
 statistics used in their scores) across all cell types or cell type pairs.}
    \item{\code{seed}}{random seed integer}
    \item{\code{trim}}{the fraction (0 to 0.5) of observations to be trimmed from each
end of x before the mean is computed. This is relevant only for the
CellPhoneDB algorithm re-implementation in liana.}
    \item{\code{parallelize}}{whether to parallelize cellphonedb-like}
    \item{\code{workers}}{number of workers to be called}
  }}
}
\value{
A list of method-resource results - i.e. provided resources are run
with each method
If only one resource is selected, a single tibble (with results for that
 resource) will be returned for each of the selected methods
}
\description{
LIANA wrapper function
}
\details{
LIANA wrapper method that can be used to call each method with
 a given set of intercellular resources from the OmniPath universe.
}
