% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_descriptive.R
\name{ligand_receptor_classes}
\alias{ligand_receptor_classes}
\title{Combines ligand-receptor and classification data}
\usage{
ligand_receptor_classes(
  ligrec,
  resource,
  attr,
  largest = NULL,
  filter_annot = TRUE,
  label_annot = identity
)
}
\arguments{
\item{ligrec}{List of tibbles with ligand-receptor data, as produced by
\code{\link{ligrec_overlap}}.}

\item{resource}{Character: name of the annotation resource.}

\item{attr}{Name of the classifying variable from the annotation resource
(e.g. pathway).}

\item{largest}{Numeric: how many of the largest groups to use. If `NULL`
all groups will be used.}

\item{filter_annot}{Expression for filtering the annotation data frame.}

\item{label_annot}{Function to process the labels of the annotation
variable (e.g. to shorten long strings).}
}
\description{
Combines ligand-receptor and classification data
}
