% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_utils.R
\docType{class}
\name{MethodSpecifics-class}
\alias{MethodSpecifics-class}
\title{S4 Class used to format benchmark output.}
\description{
S4 Class used to format benchmark output.
}
\section{Fields}{

\describe{
\item{\code{method_scores}}{}
}}

