% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natmi_pipe.R
\name{omni_to_NATMI}
\alias{omni_to_NATMI}
\title{Reform OmniPath Resource to NATMI format and save to location}
\usage{
omni_to_NATMI(op_resource, omni_path = "input/omnipath_NATMI")
}
\arguments{
\item{omni_path}{directory in which to save OP resources}

\item{resource_names}{list of omnipath resources}
}
\description{
Reform OmniPath Resource to NATMI format and save to location
}
