% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_omnipath.R
\name{omnipath_intercell}
\alias{omnipath_intercell}
\title{Retrieves intercellular interactions from OmniPath}
\usage{
omnipath_intercell(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[OmnipathR:filter_intercell_network]{OmnipathR::filter_intercell_network}}
  \describe{
    \item{\code{network}}{An intercell network data frame, as provided by
\code{\link[OmnipathR]{import_intercell_network}}, without \code{simplify}.}
    \item{\code{transmitter_topology}}{Character vector: topologies allowed for the
entities in transmitter role. Abbreviations allowed: "sec", "pmtm"
and "pmp".}
    \item{\code{receiver_topology}}{Same as \code{transmitter_topology} for the
entities in the receiver role.}
    \item{\code{min_curation_effort}}{Numeric: a minimum value of curation effort
(resource-reference pairs) for network interactions. Use zero to
disable filtering.}
    \item{\code{min_resources}}{Numeric: minimum number of resources for
interactions. The value 1 means no filtering.}
    \item{\code{min_references}}{Numeric: minimum number of references for
interactions. Use zero to disable filtering.}
    \item{\code{min_provenances}}{Numeric: minimum number of provenances (either
resources or references) for interactions. Use zero or one to
disable filtering.}
    \item{\code{consensus_percentile}}{Numeric: percentile threshold for the consensus
score of generic categories in intercell annotations. The consensus
score is the number of resources supporting the classification of an
entity into a category based on combined information of many resources.
Here you can apply a cut-off, keeping only the annotations supported
by a higher number of resources than a certain percentile of each
category. If \code{NULL} no filtering will be performed. The value is
either in the 0-1 range, or will be divided by 100 if greater than 1.
The percentiles will be calculated against the generic composite
categories and then will be applied to their resource specific
annotations and specific child categories.}
    \item{\code{loc_consensus_percentile}}{Numeric: similar to
\code{consensus_percentile} for major localizations. For example, with
a value of 50, the secreted, plasma membrane transmembrane or
peripheral attributes will be \code{TRUE} only where at least 50
percent of the resources support these.}
    \item{\code{ligand_receptor}}{Logical. If \code{TRUE}, only \emph{ligand} and
\emph{receptor} annotations will be used instead of the more generic
\emph{transmitter} and \emph{receiver} categories.}
    \item{\code{simplify}}{Logical: keep only the most often used columns. This
function combines a network data frame with two copies of the
intercell annotation data frames, all of them already having quite
some columns. With this option we keep only the names of the
interacting pair, their intercellular communication roles, and the
minimal information of the origin of both the interaction and
the annotations.}
    \item{\code{unique_pairs}}{Logical: instead of having separate rows for each
pair of annotations, drop the annotations and reduce the data frame to
unique interacting pairs. See \code{\link[OmnipathR]{unique_intercell_network}}
for details.}
    \item{\code{omnipath}}{Logical: shortcut to include the \emph{omnipath} dataset
in the interactions query.}
    \item{\code{ligrecextra}}{Logical: shortcut to include the \emph{ligrecextra}
dataset in the interactions query.}
    \item{\code{kinaseextra}}{Logical: shortcut to include the \emph{kinaseextra}
dataset in the interactions query.}
    \item{\code{pathwayextra}}{Logical: shortcut to include the \emph{pathwayextra}
dataset in the interactions query.}
  }}
}
\value{
A tibble with Intercell interactions from OmniPath
}
\description{
Retrieves intercellular interactions from OmniPath
}
