% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_omnipath.R
\name{omnipath_intercell}
\alias{omnipath_intercell}
\title{Retrieves intercellular interactions from OmniPath}
\usage{
omnipath_intercell(quality = NULL, ligrec = FALSE)
}
\value{
A tibble with Intercell interactions from OmnIPath
}
\description{
Retrieves intercellular interactions from OmniPath
}
