% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_omnipath.R
\name{omnipath_partners}
\alias{omnipath_partners}
\title{Retrieves intercellular communication partners (transmitters or receivers)
from OmniPath}
\usage{
omnipath_partners(side, quality = NULL, ligrec = FALSE)
}
\description{
Retrieves intercellular communication partners (transmitters or receivers)
from OmniPath
}
