% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_omnipath.R
\name{omnipath_partners}
\alias{omnipath_partners}
\title{Retrieves intercellular communication partners (transmitters or receivers)
from OmniPath}
\usage{
omnipath_partners(side, ...)
}
\arguments{
\item{side}{'ligand' (trans), 'receptor' (rec) or 'both' (both short or
long notation can be used)}

\item{...}{Passed to \code{OmnipathR::import_omnipath_intercell}.}
}
\description{
Retrieves intercellular communication partners (transmitters or receivers)
from OmniPath
}
