% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_descriptive.R
\name{overheat_save}
\alias{overheat_save}
\title{Save Overlap Heatmaps}
\usage{
overheat_save(df, plotname, guide_title)
}
\arguments{
\item{df}{df to be pivotted and used to plot the heatmap}

\item{plotname}{name of the plot to be saved}
}
\description{
Save Overlap Heatmaps
}
