% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_tensor.R
\name{plot_c2c_overview}
\alias{plot_c2c_overview}
\title{Function to plot an Overview of tensor-c2c results}
\usage{
plot_c2c_overview(factors)
}
\arguments{
\item{factors}{factors output from tensor-cell2cel formatted with
`format_c2c_factors`}
}
\description{
Function to plot an Overview of tensor-c2c results
}
