% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_tensor.R
\name{plot_context_boxplot}
\alias{plot_context_boxplot}
\title{Function to get boxplots with significance}
\usage{
plot_context_boxplot(contexts, test = "t_test", ...)
}
\arguments{
\item{test}{test to be performed (t_test - default) or any others from the
`rstatix` package, others include anova_test, wilcox_test, kruskal_test, etc.}

\item{...}{arguments passed to the test used.}

\item{data}{a tibble or a dataframe}
}
\description{
Function to get boxplots with significance
}
