% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_call.R
\name{recomplexify}
\alias{recomplexify}
\title{Helper function to account for complexes in the resources}
\usage{
recomplexify(lr_res, columns, protein = "complex", complex_policy = "min0")
}
\arguments{
\item{columns}{columns to account for complexes for}

\item{protein}{whether to return complexes ('complex') or protein subunits ('protein' - default)}

\item{complex_policy}{policy how to account for the presence of complexes.
Following the example of \href{https://squidpy.readthedocs.io/en/stable/api/squidpy.gr.ligrec.html}{Squidpy} valid options are:
`'min0'` (default): select the subunit with the change/expression closest to 0
 (as in \href{https://github.com/Teichlab/cellphonedb}{CellPhoneDB} when working with expression)
`'all'`: returns all subunits separately;
Alternatively, pass any other base or user-defined function that reduces a vector to a single number (e.g. mean, min, etc)}

\item{lr_cmplx}{decomplexified lr_res}
}
\value{
complex-accounted lr_res
}
\description{
Helper function to account for complexes in the resources
}
\details{
to be passed before the relevant score_calc function
}
