% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_call.R
\name{recomplexify}
\alias{recomplexify}
\title{Helper function to account for complexes in the resources}
\usage{
recomplexify(lr_res, columns, protein = "complex", complex_policy = "min0")
}
\arguments{
\item{columns}{columns to account for complexes for}

\item{complex_policy}{policy how to account for the presence of complexes.
Following the example of \url{https://squidpy.readthedocs.io/en/stable/api/squidpy.gr.ligrec.html}{Squidpy} valid options are:
'min0' select the subunit with the change/expression closest to 0 (as in CellPhoneDB)
'all' returns all subunits separately}

\item{lr_cmplx}{decomplexified lr_res}
}
\value{
complex-accounted lr_res
}
\description{
Helper function to account for complexes in the resources
}
\details{
to be passed before the relevant score_calc function
}
