% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_omnipath.R
\name{reform_omni}
\alias{reform_omni}
\title{Helper Function to Reformat ligrec for LR Pipeline}
\usage{
reform_omni(ligrec)
}
\arguments{
\item{ligrec}{OmniPath list returned by compile_ligrec}
}
\value{
A list of OmniPath resources, including OmniPath composite DB,
A reshuffled OmniPath, and a Default with NULL ( tool pipelines run
using their default resource)
}
\description{
Helper Function to Reformat ligrec for LR Pipeline
}
