% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liana_wrapper.R
\name{select_resource}
\alias{select_resource}
\title{Helper Function to Handle resource choices}
\usage{
select_resource(resource)
}
\arguments{
\item{resource}{names of the resources}
}
\description{
Helper Function to Handle resource choices
}
