% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_omnipath.R
\name{shuffle_omnipath}
\alias{shuffle_omnipath}
\title{Shuffle OmniPath Intercell DB}
\usage{
shuffle_omnipath(op_resource, .seed = 1004)
}
\arguments{
\item{op_resource}{Intrcell DB to shuffle}

\item{.seed}{Value for set.seed}
}
\value{
A shuffled omnipath-formatted resource
}
\description{
Shuffle OmniPath Intercell DB
}
