% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_crc_data.R
\name{sparsify_to_seurat}
\alias{sparsify_to_seurat}
\title{Helper function to convert CRC data to sparse Seurat}
\usage{
sparsify_to_seurat(counts_loc, meta_loc, save_loc)
}
\arguments{
\item{counts_loc}{counts location}

\item{meta_loc}{metadata location}

\item{save_loc}{location to save sparsified Seurat object}
}
\value{
Return a Sparse Seurat Object
}
\description{
Helper function to convert CRC data to sparse Seurat
}
